% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asana_api.R
\name{add_asana_tag_to_task}
\alias{add_asana_tag_to_task}
\title{add_asana_tag_to_task}
\usage{
add_asana_tag_to_task(task_gid, tag_gid, api_token, logger)
}
\arguments{
\item{task_gid}{The GID of the task}

\item{tag_gid}{The GID of the tag to add}

\item{api_token}{The Asana API token}

\item{logger}{The logger object}
}
\value{
TRUE if the tag was added successfully, FALSE otherwise
}
\description{
Adds a tag to an Asana task
}
