% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_api_contact_details.R
\name{add_contact_details}
\alias{add_contact_details}
\title{add_contact_details}
\usage{
add_contact_details(headers, df)
}
\arguments{
\item{headers}{the header informations you have to send with your request}

\item{df}{the dataframe which should include the following fields:}
}
\value{
no return values
}
\description{
This function calls the CRM API and starts a post request for new contact details,
attached to companies or persons.
attachable_id - id of the person / company
action - has to be value "add"
attachable_type - "people" or "companies"
field_name - the contact detail value (e.g., email address or phone number)
atype - type of contact detail (e.g., "office", "work", "mobile")
contact_detail_type - type of resource ("email", "tel", "homepage", "sm" etc.)
field_type - has to be value "contact_details"
}
