% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_api_tags.R
\name{add_crm_tag}
\alias{add_crm_tag}
\title{add_crm_tag}
\usage{
add_crm_tag(headers, df)
}
\arguments{
\item{headers}{the header informations you have to send with your request}

\item{df}{the dataframe which should include the following fields:}
}
\value{
no return values
}
\description{
This function calls the CRM API and starts a post request for new tags,
attached to companies or persons.
attachable_id - id of the person / company
custom_fields_id - optional, id of the current tag
action - optional, has to be value "add"
attachable_type - "people" or "companies"
field_name - name of the tag
field_type - optional, has to be value "tag"
}
