% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_export_to_jp5.R
\name{aggregate_export_jp5}
\alias{aggregate_export_jp5}
\title{aggregation for excel files for export to jp5}
\usage{
aggregate_export_jp5(
  tmp_folder = "../../base-data/PMI/export/tmp/",
  output_folder = "../../base-data/PMI/export/"
)
}
\arguments{
\item{tmp_folder}{the location where the temporary files are stored}

\item{export_folder}{the location where the final combined file is stored}
}
\value{
this fcuntion returns nothing, but has side effects.
}
\description{
The function compiles excel files in a given directory and returns a large excelfile
}
