% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{analyze_postgres_tables_file}
\alias{analyze_postgres_tables_file}
\title{analyze_postgres_tables_file}
\usage{
analyze_postgres_tables_file()
}
\description{
Extract postgres tables from the entire current R script.
}
