% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{analyze_postgres_tables_project}
\alias{analyze_postgres_tables_project}
\title{analyze_postgres_tables_project}
\usage{
analyze_postgres_tables_project()
}
\description{
Extract postgres tables from all R scripts in current project, grouped by file.
}
