% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{analyze_postgres_tables_selection}
\alias{analyze_postgres_tables_selection}
\title{analyze_postgres_tables_selection}
\usage{
analyze_postgres_tables_selection()
}
\description{
Extract postgres tables from the current RStudio selection
}
