% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_authentication.R
\name{authentication_personio}
\alias{authentication_personio}
\title{authentication_personio}
\usage{
authentication_personio(args)
}
\arguments{
\item{args}{Additional input parameter, only needed through FlowForce Job}
}
\value{
Named list containing decrypted client_id, client_secret, and access_token
}
\description{
This function handles the authentication for the Personio API.
It decrypts the client_id and client_secret, then requests an access token from the Personio API.
It supports manual decryption key input as well as FlowForce arguments.
}
