% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asana_api.R
\name{check_asana_tag_exists_in_workspace}
\alias{check_asana_tag_exists_in_workspace}
\title{check_asana_tag_exists_in_workspace}
\usage{
check_asana_tag_exists_in_workspace(tag_gid, api_token, logger)
}
\arguments{
\item{tag_gid}{The GID of the tag to check}

\item{api_token}{The Asana API token}

\item{logger}{The logger object}
}
\value{
TRUE if the tag exists, FALSE if not, or NA for an error
}
\description{
Checks if a tag exists in a workspace
}
