% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contracts_to_bill.R
\name{check_contract_starts}
\alias{check_contract_starts}
\title{check_contract_starts}
\usage{
check_contract_starts(
  df_current_contracts,
  df_reale_starts,
  this_period,
  planned_start = TRUE
)
}
\arguments{
\item{df_current_contracts}{The df with the contracts, each contract and product has on row}

\item{this_period}{this should be a lubridate interval in which the period of starts is given}

\item{planned_start}{when true it returns campaigns based on the contracted start date, when false for the real starting date}

\item{reale_starts_df}{The df which contains the real running times, given the campaigns started}
}
\value{
The function returns the all the contracts that started
}
\description{
check_contract_starts
}
