% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_information_billomat_functions.R
\name{compress_column}
\alias{compress_column}
\title{compress column}
\usage{
compress_column(column)
}
\arguments{
\item{column}{the path to the billomat db}
}
\value{
the string
}
\description{
this function shortens names.
}
