% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenue_functions.R
\name{compute_revenue_month}
\alias{compute_revenue_month}
\title{compute_revenue_month}
\usage{
compute_revenue_month(df)
}
\arguments{
\item{df}{the data frame which must contain the customer ("Kunde"), confirmation_number, Month and MonthlyRevenue}
}
\value{
returns a tibble with grouped revenue streams
}
\description{
Here I compute the revenue per month considering the booking date and the sales person.
Function uses the revenue_month() function to purrr over the df creating the revenue types
}
