% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenue_functions.R
\name{compute_yearly_nrr_month}
\alias{compute_yearly_nrr_month}
\title{compute_yearly_nrr_month}
\usage{
compute_yearly_nrr_month(
  df,
  focus_company = FALSE,
  focus_year = TRUE,
  pasting = FALSE,
  pasting_p = 0.7
)
}
\arguments{
\item{df}{contains the data frame with all contracts}

\item{focus_company}{is set to FALSE because I want to compute the overall NRR}

\item{focus_year}{is set to to TRUE because I want to compute the overall NRR for the year}

\item{pasting}{is set to FALSE because I want THE number and not a sampled estimate}

\item{pasting_p}{is the proportion of companies that should be sampled - if pasting is TRUE,
then this is relevant for setting the sampling proportion in the pasting}
}
\value{
this returns a tibble with the main KPIs as columns per year
}
\description{
The function is used to purrr over all contracts and get the KPI of customers, type of revenue, churn,
nrr and dbner per year.
}
