% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_information_billomat_functions.R
\name{consolidate_invoice_information}
\alias{consolidate_invoice_information}
\title{consolidate_invoice_information
This function is used to create one dataframe from the billing information in the comments
and the billing information from the document. The information from the comments supersede the information
in the document.}
\usage{
consolidate_invoice_information(
  df_Billing_information_comment,
  df_Billing_information_document
)
}
\arguments{
\item{df_Billing_information_comment}{the dataframe containing the extracted information of comments}

\item{df_Billing_information_document}{the dataframe containing the extracted information of the document}
}
\value{
The function returns one dataframe with the document_id, key col containing the name of the information and value its value.
}
\description{
consolidate_invoice_information
This function is used to create one dataframe from the billing information in the comments
and the billing information from the document. The information from the comments supersede the information
in the document.
}
