% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{convert_list_to_tibble}
\alias{convert_list_to_tibble}
\title{convert_list_to_tibble
General function to convert each contact into a tibble row}
\usage{
convert_list_to_tibble(df_list)
}
\arguments{
\item{df_list}{a list where each entry is its own list.}
}
\value{
the call returns a tibble
}
\description{
convert_list_to_tibble
General function to convert each contact into a tibble row
}
