% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asana_api.R
\name{create_asana_subtask}
\alias{create_asana_subtask}
\title{create_asana_subtask}
\usage{
create_asana_subtask(
  parent_task_gid,
  subtask_name,
  api_token,
  logger,
  assignee = NULL,
  due_on = NULL,
  notes = NULL
)
}
\arguments{
\item{parent_task_gid}{The GID of the parent task where the subtask will be created}

\item{subtask_name}{The name of the subtask}

\item{api_token}{Your Asana API token}

\item{logger}{The logger object}

\item{assignee}{(Optional) The GID of the assignee (default: NULL)}

\item{due_on}{(Optional) The due date in "YYYY-MM-DD" format (default: NULL)}

\item{notes}{(Optional) Notes or description for the subtask (default: NULL)}
}
\value{
The GID of the created subtask or NULL if the creation failed
}
\description{
Creates a subtask in an Asana task and returns its GID
}
