% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asana_api.R
\name{create_asana_tag}
\alias{create_asana_tag}
\title{create_asana_tag}
\usage{
create_asana_tag(workspace_id, tag_name, asana_token, logger)
}
\arguments{
\item{workspace_id}{The ID of the workspace where the tag will be created}

\item{tag_name}{The name of the new tag}

\item{asana_token}{Personal Access Token (PAT) for Asana API authentication}

\item{logger}{The logger object}
}
\value{
The ID (GID) of the newly created tag or NULL if creation fails
}
\description{
Creates a new tag in Asana and returns the tag ID
}
