% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{create_contacts_tibble}
\alias{create_contacts_tibble}
\title{create_contacts_tibble
Function takes the list wise returned values from the api call to get contacts and
turns them into a tibble}
\usage{
create_contacts_tibble(df_list)
}
\arguments{
\item{df_list}{a list where each entry is its own list.}
}
\value{
the call returns a tibble
}
\description{
create_contacts_tibble
Function takes the list wise returned values from the api call to get contacts and
turns them into a tibble
}
