% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_api_people.R
\name{create_crm_person}
\alias{create_crm_person}
\title{create_person}
\usage{
create_crm_person(headers, df)
}
\arguments{
\item{headers}{API headers with authentication}

\item{df}{Dataframe with person information}
}
\value{
no return values
}
\description{
Create new person in CRM
Required columns: action, field_type
Optional columns: salutation, first_name, last_name, background
action must be "create", field_type must be "person"
}
