% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{create_entries_for_all_pages}
\alias{create_entries_for_all_pages}
\title{create_entries_for_all_pages}
\usage{
create_entries_for_all_pages(data)
}
\arguments{
\item{data}{are all pages that are retrieved as xml}
}
\description{
this function uses the create_entries_for_page an purrrs over all pages
}
