% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{create_entries_for_page}
\alias{create_entries_for_page}
\title{create_entries_for_page}
\usage{
create_entries_for_page(entry_as_xml)
}
\arguments{
\item{entry_as_xml}{is the data of one entry as extract_single_entry_client_properties}
}
\description{
this function uses
}
