% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_crm_functions.R
\name{create_full_table}
\alias{create_full_table}
\title{create_full_table
this function joins all retrieved lists together in one long table
all_lists <- c("emails","tags","positions","companies","tels","tasks","tasks_pending")}
\usage{
create_full_table(df, all_lists)
}
\arguments{
\item{df}{the df which contains the list wise variable}

\item{all_lists}{the lists which should be transformed -- this has to be a list with this name}
}
\value{
the tibble which contains all information, this is stored in a tibble
to create the dataset, run the function and assign the result to a named object:
result <- create_full_table(crm_data,all_lists)
The id of the person is the primary key to link the resulting tables together
}
\description{
create_full_table
this function joins all retrieved lists together in one long table
all_lists <- c("emails","tags","positions","companies","tels","tasks","tasks_pending")
}
