% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fields_main_document.R
\name{create_header_billing_info_text}
\alias{create_header_billing_info_text}
\title{create_header_billing_info_text}
\usage{
create_header_billing_info_text(df_positions, df_information_bill)
}
\arguments{
\item{df_positions}{the positions that need to get the billing information attached to}

\item{df_information_bill}{The df containing the billing information after comments and intro/ note were consolidated}
}
\value{
The function returns the information added to the fields and shortened to 50 characters
}
\description{
this function uses are table with all fields for billing information and creates a wide table to show it on the billing doc
}
