% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_information_billomat_functions.R
\name{create_laufzeiten_confirmations}
\alias{create_laufzeiten_confirmations}
\title{extract the information from each note and put it in a long table
create the final dataframe from the provided data -- important, names must match
the data provided is split with a separate symbol}
\usage{
create_laufzeiten_confirmations(df, confirmation_items, confirmations)
}
\arguments{
\item{df}{the df which holds the data from which information is extracted}

\item{confirmation_items}{the confirmation items}

\item{confirmations}{the confirmations}
}
\value{
the data as df which has id and all information
}
\description{
extract the information from each note and put it in a long table
create the final dataframe from the provided data -- important, names must match
the data provided is split with a separate symbol
}
