% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{create_new_database}
\alias{create_new_database}
\title{create_new_database}
\usage{
create_new_database(billomatDB, content, data_db, keys_db)
}
\arguments{
\item{billomatDB}{is an DBI object containing the path to the db}

\item{content}{is the name of the table}

\item{data_db}{the tibble containing the data}

\item{keys_db}{the encryption key for the database}
}
\description{
this function deletes the db if it exists and adds the new data as new table.
}
