% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_data_processing.R
\name{create_single_table}
\alias{create_single_table}
\title{create_single_table
this function creates the a single crm table with all information as a tibble from one list
the table is long and contains a name and value for each id
to ensure that names can be distinct, I create an entry_number based on the grouped id and nam.
If an id has the same field multiple times, then the entry_number will be >1.}
\usage{
create_single_table(df, listname = "companies", i = 1)
}
\arguments{
\item{df}{the df which contains the list wise variable}

\item{listname}{the lists which should be transformed -- this has to be a list with this name}
}
\value{
the tibble which contains all information, this is stored in a tibble
}
\description{
create_single_table
this function creates the a single crm table with all information as a tibble from one list
the table is long and contains a name and value for each id
to ensure that names can be distinct, I create an entry_number based on the grouped id and nam.
If an id has the same field multiple times, then the entry_number will be >1.
}
