% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_request.R
\name{crm_DELETE}
\alias{crm_DELETE}
\title{CRM DELETE Request with Rate Limit Retry}
\usage{
crm_DELETE(url, headers, encode = "raw")
}
\arguments{
\item{url}{API-Endpunkt URL}

\item{headers}{Request-Headers}

\item{encode}{Encoding (default: "raw")}
}
\value{
httr response
}
\description{
CRM DELETE Request with Rate Limit Retry
}
\keyword{internal}
