% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_request.R
\name{crm_GET}
\alias{crm_GET}
\title{CRM GET Request with Rate Limit Retry}
\usage{
crm_GET(url, headers)
}
\arguments{
\item{url}{API-Endpunkt URL}

\item{headers}{Request-Headers}
}
\value{
httr response
}
\description{
CRM GET Request with Rate Limit Retry
}
\keyword{internal}
