% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_request.R
\name{crm_POST}
\alias{crm_POST}
\title{CRM POST Request with Rate Limit Retry}
\usage{
crm_POST(url, headers, body = NULL, encode = "json")
}
\arguments{
\item{url}{API-Endpunkt URL}

\item{headers}{Request-Headers}

\item{body}{Request-Body}

\item{encode}{Encoding (default: "json")}
}
\value{
httr response
}
\description{
CRM POST Request with Rate Limit Retry
}
