% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{curl_fetch_billomat}
\alias{curl_fetch_billomat}
\title{curl_fetch_billomat
this function carries out the call to get the header}
\usage{
curl_fetch_billomat(
  content = c("articles", "clients", "offers", "invoices", "confirmations"),
  page = 1,
  per_page = per_page,
  billomatApiKey = billomatApiKey,
  billomatID = billomatID
)
}
\arguments{
\item{content}{the name of the tables you are interested in, this is one string}

\item{page}{how many entries per page to get}

\item{billomatApiKey}{please provide your billomat Api key here}

\item{billomatID}{please provide your billomat ID here}
}
\value{
the call returns a list with the header information
}
\description{
curl_fetch_billomat
this function carries out the call to get the header
}
