% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_aggregate_data_function.R
\name{custom_aggregate_data}
\alias{custom_aggregate_data}
\title{custom_aggregate_data}
\usage{
custom_aggregate_data(
  df,
  Monat,
  Category_Column = "Produkt_short",
  Start_Date,
  End_Date,
  turn2wide = FALSE,
  aggregate_function = "sum",
  dependent_variable = "MonthlyRevenue"
)
}
\arguments{
\item{df}{the data.frame to aggregate over}

\item{Monat}{choose Date Column to calculate Monat upon; "Auftragsmonat" for order month or "Month" for booking month of earnings}

\item{Category_Column}{choose column with factors to aggregate by; Default is "Product_short" to aggregate by Studyflixs' products}

\item{Start_Date}{Start date to start aggregating data in data.frame}

\item{End_Date}{End date to end aggregating data in data.frame}

\item{aggregate_function}{function to aggregate data by; Default is "sum"; the other possible aggregate_function is "count"}

\item{turn_2_wide}{boolean wether to plot data or show as table; TRUE: pivots table, use to show summary table; FALSE: default, use to generate plot with aggregated data}

\item{dependenet_variable}{column in df to apply aggregate_function to; Default is "MonthlyRevenue"}
}
\value{
the aggregated data.frame indexed by months; if turn_2_wide = TRUE then Categories in columns and column "Summe" is added to the right of the aggregated data.frame
}
\description{
This function aggregates monthly data across a data.frame by applying an aggregate_function and respecting Categories.
It can be either used to aggregate data for a bar chart or display as an aggregated table.
}
