% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{custom_decrypt_db}
\alias{custom_decrypt_db}
\title{custom_decrypt_db}
\usage{
custom_decrypt_db(df, columns_to_decrypt, key = NULL)
}
\arguments{
\item{df}{The dataframe to be decrypted.}

\item{columns_to_decrypt}{The columns that need to be decrpyted.}

\item{key}{The key.}
}
\value{
The specified columns of the database table gets decrypted
}
\description{
This function is used to decrypt a dataframe that is saved within a sqlite database.
You are able to specifiy which columns you want to decrypt.
}
