% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asana_api.R
\name{delete_asana_tag}
\alias{delete_asana_tag}
\title{delete_asana_tag}
\usage{
delete_asana_tag(tag_id, asana_key, request_func = DELETE)
}
\arguments{
\item{tag_id}{Tag id to delete}

\item{asana_key}{Asana key (bearer) for the asana board}
}
\value{
Logical. TRUE on success, FALSE on failure
}
\description{
Deletes an Asana tag
}
