% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_api_people.R
\name{delete_crm_person}
\alias{delete_crm_person}
\title{delete_crm_person}
\usage{
delete_crm_person(headers, df)
}
\arguments{
\item{headers}{API headers with authentication (including X-apikey)}

\item{df}{Dataframe with person information to delete}
}
\value{
No return value (invisible NULL)
}
\description{
This function calls the CRM API and deletes a person by their ID.
Required columns: attachable_id, action, field_type
action must be "delete", field_type must be "person"
}
