% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenue_functions.R
\name{df_transpose}
\alias{df_transpose}
\title{df_transpose}
\usage{
df_transpose(df)
}
\arguments{
\item{df}{the tabel to transpose}
}
\value{
a tibble
}
\description{
this function transposes a table
}
