% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{download_all_tables}
\alias{download_all_tables}
\title{download_all_tables
this function download pulls the selected tables from the Billomat API}
\usage{
download_all_tables(content)
}
\arguments{
\item{content}{the name of the entry in an xml you are interested in}
}
\value{
the call returns a dataframe which contains the id of the entry and all information in one long df
}
\description{
download_all_tables
this function download pulls the selected tables from the Billomat API
}
