% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsgvo_suppression.R
\name{dsgvo_email_tombstone}
\alias{dsgvo_email_tombstone}
\title{Personen-stabiler Tombstone aus einem E-Mail-Hash}
\usage{
dsgvo_email_tombstone(email_hash)
}
\arguments{
\item{email_hash}{Gepfefferter SHA-256 der E-Mail.}
}
\value{
Tombstone-String \verb{[geloescht]-<hash>}.
}
\description{
Personen-stabiler Tombstone aus einem E-Mail-Hash
}
