% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsgvo_suppression.R
\name{dsgvo_parse_pg_array}
\alias{dsgvo_parse_pg_array}
\title{Parst ein PG-Array-Literal ('{a,b}') zu einem Character-Vektor (leere/NA -> character(0))}
\usage{
dsgvo_parse_pg_array(s)
}
\arguments{
\item{s}{Ein '{...}'-String ODER bereits ein Vektor.}
}
\value{
Character-Vektor.
}
\description{
Parst ein PG-Array-Literal ('{a,b}') zu einem Character-Vektor (leere/NA -> character(0))
}
\keyword{internal}
