% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_process_attachments.R
\name{enrich_offers_ABs_with_start_end}
\alias{enrich_offers_ABs_with_start_end}
\title{enrich_offers_ABs_with_start_end}
\usage{
enrich_offers_ABs_with_start_end(crm_ABs_Offers, latest_rev_per_mensem)
}
\arguments{
\item{crm_ABs_Offers}{data frame from matching_ABs_Billomat_CRM}

\item{latest_rev_per_mensem}{rev_per_mensem}
}
\value{
enriched crm_ABs_Offers with start and end of campaign
}
\description{
enrich_offers_ABs_with_start_end
}
