% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{extract_all_table_names}
\alias{extract_all_table_names}
\title{Extract All Table Names from a Lazy Query Object}
\usage{
extract_all_table_names(x)
}
\arguments{
\item{x}{Das \code{lazy_query}-Objekt eines \code{tbl}.}
}
\value{
Ein Character-Vektor mit eindeutigen Tabellennamen.
}
\description{
Rekursiv alle Tabellen-Namen aus einem \code{lazy_query}-Objekt extrahieren.
}
\keyword{internal}
