% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_information_billomat_functions.R
\name{extract_laufzeit_dauer}
\alias{extract_laufzeit_dauer}
\title{extract the laufzeit duration information and put it into a table with id, entrynumber and laufzeit}
\usage{
extract_laufzeit_dauer(df_position, field = "description")
}
\arguments{
\item{df_position}{the df which holds the data from which information is extracted}

\item{field}{the field name as string from which the data is extracted}
}
\value{
the data as df with the id, entry number per position and the extracted laufzeit
only the first entry should be used, multiple entries indicate that multiple products are in the position
}
\description{
extract the laufzeit duration information and put it into a table with id, entrynumber and laufzeit
}
