% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_information_billomat_functions.R
\name{extract_note}
\alias{extract_note}
\title{create a readable concise note}
\usage{
extract_note(df, field)
}
\arguments{
\item{df}{the df which holds the data from which information is extracted}

\item{field}{the name of the field that contains the data}
}
\value{
the data as df which has id and all information
}
\description{
This function creates a readable note
}
