% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_process_attachments.R
\name{extract_offers_and_abs_from_crm}
\alias{extract_offers_and_abs_from_crm}
\title{matching_table_AB_AN}
\usage{
extract_offers_and_abs_from_crm(attachments)
}
\arguments{
\item{attachments}{all attachments from CRM}
}
\value{
df with attachment creation
}
\description{
matching_table_AB_AN
}
