% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{extract_postgres_tables}
\alias{extract_postgres_tables}
\title{Extract postgres db tables from text}
\usage{
extract_postgres_tables(text)
}
\arguments{
\item{text}{character, code to parse}
}
\value{
character vector of table names
}
\description{
Extract postgres db tables from text
}
