% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{extract_single_entry}
\alias{extract_single_entry}
\title{extract_single_entry
this function extracts a single entry from an xml}
\usage{
extract_single_entry(entry_as_xml)
}
\arguments{
\item{entry_as_xml}{the name of the entry in an xml you are interested in}
}
\value{
the call returns a dataframe which contains the id of the entry and all information in one long df
}
\description{
extract_single_entry
this function extracts a single entry from an xml
}
