% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{extract_single_entry_client_properties}
\alias{extract_single_entry_client_properties}
\title{extract_single_entry_client_properties}
\usage{
extract_single_entry_client_properties(entry_as_xml)
}
\arguments{
\item{entry_as_xml}{is the data of one entry as an xml}
}
\description{
this function extracts the properties of a single entry
}
