% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{extract_xml}
\alias{extract_xml}
\title{extract_xml
this function extracts a all entries from an xml}
\usage{
extract_xml(data, list_num)
}
\arguments{
\item{data}{the name of the entry in an xml you are interested in}

\item{list_num}{the index ot the list which contains the data}
}
\value{
the call returns a dataframe which contains the id of the entry and all information in one long df
}
\description{
extract_xml
this function extracts a all entries from an xml
}
