% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{fetch_all_entries}
\alias{fetch_all_entries}
\title{fetch_all_entries
this function carries out the call to get all the content of the call
This is a new function using httr2}
\usage{
fetch_all_entries(billomatID, content, billomatApiKey, per_page = 100)
}
\arguments{
\item{billomatID}{please provide your billomat ID here}

\item{content}{the name of the tables you are interested in, this is one string}

\item{billomatApiKey}{please provide your billomat Api key here}

\item{per_page}{how many entries per page to get}
}
\value{
the call returns a list for each page
}
\description{
fetch_all_entries
this function carries out the call to get all the content of the call
This is a new function using httr2
}
