% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{fetch_header_httr2}
\alias{fetch_header_httr2}
\title{fetch_header_httr2}
\usage{
fetch_header_httr2(
  content = c("articles", "clients", "offers", "invoices", "confirmations"),
  billomatApiKey = billomatApiKey,
  billomatID = billomatID
)
}
\arguments{
\item{content}{the name of the tables you are interested in, this is one string}

\item{billomatApiKey}{please provide your billomat Api key here}

\item{billomatID}{please provide your billomat ID here}

\item{page}{how many entries per page to get}
}
\value{
the call returns a list with the header information
}
\description{
this function carries out the call to get the header
}
