% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_helpers.R
\name{filter_by_field_and_action}
\alias{filter_by_field_and_action}
\title{Filter dataframe by field_type and action}
\usage{
filter_by_field_and_action(df, field_type_value, action_value)
}
\arguments{
\item{df}{dataframe to filter}

\item{field_type_value}{value to filter field_type by (e.g., "tag", "custom_field")}

\item{action_value}{value to filter action by (e.g., "add", "remove")}
}
\description{
Filter dataframe by field_type and action
}
\keyword{internal}
