% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_positions_to_bill.R
\name{filter_positions2bill}
\alias{filter_positions2bill}
\title{filter_positions2bill}
\usage{
filter_positions2bill(sap_posten_confirmation, all_positions)
}
\arguments{
\item{sap_posten_confirmation}{The dataframe containing the amounts that have been billed}

\item{all_positions}{The dataframe containing all positions}
}
\value{
The function returns a dataframe containing all positions to bill
}
\description{
filter_positions2bill
}
