% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_information_billomat_functions.R
\name{fix_invalid_dmy_vec}
\alias{fix_invalid_dmy_vec}
\title{Fix Invalid Dates in "yyyy-mm-dd" Format}
\usage{
fix_invalid_dmy_vec(date_strs)
}
\arguments{
\item{date_strs}{A character vector of dates in the "yyyy-mm-dd" format.}
}
\value{
A vector of valid \code{Date} objects.
}
\description{
This function takes a character vector of dates in "yyyy-mm-dd" format and
corrects invalid dates by adjusting the day to the last valid day of the month.
For example, "2025-02-29" becomes "2025-02-28" since 2025 is not a leap year.
}
\examples{
fix_invalid_ymd_vec(c("2025-02-29", "2024-11-31", "2025-05-15"))
# Returns: "2025-02-28", "2024-11-30", "2025-05-15"

}
