% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.R
\name{generate_password}
\alias{generate_password}
\title{generate_password}
\usage{
generate_password(length = 10)
}
\arguments{
\item{length}{the length of the password you want to generate}
}
\value{
a password including lower and upper case letters, numbers and special characters
}
\description{
This function creates a random password of a specified length using lowercase
letters, uppercase letters, numbers, and special characters, ensuring at
least one character from each set.
}
