% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asana_api.R
\name{get_all_asana_tags}
\alias{get_all_asana_tags}
\title{get_all_asana_tags}
\usage{
get_all_asana_tags(workspace_id, asana_key)
}
\arguments{
\item{workspace_id}{Workspace id for the asana board}

\item{asana_key}{Asana key (bearer) for the asana board}
}
\value{
All tags in the asana board
}
\description{
Gets all tags from the Asana board
}
