% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{get_all_client_properties}
\alias{get_all_client_properties}
\title{get_all_client_properties}
\usage{
get_all_client_properties(
  content = "`client-property-values`",
  billomatApiKey = billomatApiKey,
  billomatID = billomatID,
  billomatDB = billomatDB,
  encryption_key_db = encryption_db
)
}
\arguments{
\item{content}{the name of the tables you are interested in, this is one string}

\item{billomatApiKey}{please provide your billomat Api key here}

\item{billomatID}{please provide your billomat ID here}

\item{per_page}{how many entries per page to get}
}
\description{
this function downloads all client properties and saves them in the billomat db
}
