% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asana_api.R
\name{get_asana_custom_field_gids}
\alias{get_asana_custom_field_gids}
\title{get_asana_custom_field_gids}
\usage{
get_asana_custom_field_gids(project_id, api_token, logger)
}
\arguments{
\item{project_id}{The GID of the project}

\item{api_token}{The Asana API token}

\item{logger}{The logger object}
}
\value{
A list of custom field GIDs
}
\description{
Gets the gids of all custom fields for a project
}
