% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asana_api.R
\name{get_asana_project_sections}
\alias{get_asana_project_sections}
\title{get_asana_project_sections}
\usage{
get_asana_project_sections(project_id, api_token, logger = NULL)
}
\arguments{
\item{project_id}{The GID of the Asana project.}

\item{api_token}{Personal Access Token for Asana API authentication.}

\item{logger}{Logger object (optional).}
}
\value{
A data frame with section names and their GIDs.
}
\description{
Retrieves all sections and their GIDs for a given Asana project
}
