% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asana_api.R
\name{get_asana_subtasks}
\alias{get_asana_subtasks}
\title{get_asana_subtasks}
\usage{
get_asana_subtasks(task_id, api_token, logger)
}
\arguments{
\item{task_id}{The Asana task ID}

\item{api_token}{The Asana API token}

\item{logger}{The logger object}
}
\value{
A list containing subtask details or NULL if an error occurs
}
\description{
Retrieves the subtasks of a given Asana task (httr2 version)
}
