% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asana_api.R
\name{get_asana_task_assignee}
\alias{get_asana_task_assignee}
\title{get_asana_task_assignee}
\usage{
get_asana_task_assignee(task_id, api_token, logger)
}
\arguments{
\item{task_id}{Character. The GID of the Asana task.}

\item{api_token}{Character. Your Asana API token.}

\item{logger}{The logger object.}
}
\value{
Assignee details if the task has an assignee, NULL otherwise.
}
\description{
Gets the assignee of an Asana task, if it exists, NULL otherwise
}
