% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asana_api.R
\name{get_asana_task_assignee_from_url}
\alias{get_asana_task_assignee_from_url}
\title{get_asana_task_assignee_from_url}
\usage{
get_asana_task_assignee_from_url(api_token, task_url, logger)
}
\arguments{
\item{api_token}{The Asana API token}

\item{task_url}{The URL of the Asana task}

\item{logger}{The logger object}
}
\value{
The assignee information or NULL if unassigned
}
\description{
Gets the assignee of an Asana task
}
