% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asana_api.R
\name{get_asana_task_details}
\alias{get_asana_task_details}
\title{get_asana_task_details}
\usage{
get_asana_task_details(task_id, api_token, logger)
}
\arguments{
\item{task_id}{The Asana task ID}

\item{api_token}{The Asana API token}

\item{logger}{The logger object}
}
\value{
A list containing task details or NULL
}
\description{
Retrieves details of a task from Asana
}
