% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asana_api.R
\name{get_asana_task_template_details_with_nested_subtasks}
\alias{get_asana_task_template_details_with_nested_subtasks}
\title{get_asana_task_template_details_with_nested_subtasks}
\usage{
get_asana_task_template_details_with_nested_subtasks(
  task_template_gid,
  asana_token,
  logger
)
}
\arguments{
\item{task_template_gid}{The GID of the task template}

\item{asana_token}{The Asana API token}

\item{logger}{The logger object}
}
\value{
A list containing the task template details or NULL
}
\description{
Get the details of a task template in Asana, including deeper levels of subtasks
}
