% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asana_api.R
\name{get_asana_task_templates}
\alias{get_asana_task_templates}
\title{get_asana_task_templates}
\usage{
get_asana_task_templates(project_id, asana_token, logger)
}
\arguments{
\item{project_id}{The GID of the project}

\item{logger}{The logger object}
}
\value{
task templates of the project
}
\description{
Gets all tasks templates of a project
}
