% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_api_protocols.R
\name{get_central_station_attachments}
\alias{get_central_station_attachments}
\title{get_central_station_attachments}
\usage{
get_central_station_attachments(api_key, protocols_vector)
}
\arguments{
\item{api_key}{the api key you have to provide}

\item{protocols_vector}{vector with protocols, where we want to export the attachments from}
}
\value{
the tibble which contains all information, this can be stored in a single vector or lists
}
\description{
this function calls the crm api and downloads the data of attachments for a
given vector of protocol-ids
}
