% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_crm_functions.R
\name{get_central_station_cal_events}
\alias{get_central_station_cal_events}
\title{get_central_station_cal_events}
\usage{
get_central_station_cal_events(api_key, pages = 1000)
}
\arguments{
\item{api_key}{the api key you have to provide}

\item{pages}{the setting to set a maximum number of pages to download, with 1000 pages as default (one page includes 250 entries)}
}
\value{
the tibble which contains all information of every member at the cal_events
}
\description{
this function calls the crm api and downloads the cal_events (past and future) to all contact.
}
