% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_functions.R
\name{get_central_station_contacts}
\alias{get_central_station_contacts}
\title{get_central_station_contacts}
\usage{
get_central_station_contacts(api_key, pages = "all")
}
\arguments{
\item{api_key}{the api key you have to provide}

\item{pages}{the setting to call all pages (default), or a specific number of pages (one page includes 250 entries)}
}
\value{
the tibble which contains all information, this can be stored in a single vector or lists
}
\description{
this function calls the crm api and downloads all fields that belong to a contact.
The notes and protocols are NOT downloaded.
}
