% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_api_protocols.R
\name{get_central_station_protocols}
\alias{get_central_station_protocols}
\title{get_central_station_protocols}
\usage{
get_central_station_protocols(
  api_key,
  filter_by = FALSE,
  filter_vector = NA,
  pages = 2000
)
}
\arguments{
\item{api_key}{the api key you have to provide}

\item{filter_by}{setting to prefilter export by a list of "company_id" or "person_id", if all protocols are needed use FALSE}

\item{filter_vector}{vector to filter by "company_id" or "person_id" (decided by filter_by), not needed if all protocols should get downloaded}

\item{pages}{the setting to set a maximum number of pages to download, with 2000 pages as default (one page includes 250 entries)}
}
\value{
the tibble which contains all information, this can be stored in a single vector or lists
}
\description{
this function calls the crm api and downloads the protocols (notices and mails) to all contact.
}
