% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_crm_functions.R
\name{get_central_station_single_tags}
\alias{get_central_station_single_tags}
\title{get_central_station_cal_events}
\usage{
get_central_station_single_tags(api_key, tag_to_export, pages = 1000)
}
\arguments{
\item{api_key}{the api key you have to provide}

\item{tag_to_export}{the setting to set a tag-string for which all person_ids should be loaded}

\item{pages}{the setting to set a maximum number of pages to download, with 1000 pages as default (one page includes 250 entries)}
}
\value{
the tibble which contains all person_ids with the defined tag, the dates when it was created and updated and the method and account who assigned it
}
\description{
this function calls the crm api and downloads the all person_ids with one defined tag and the time it was set.
}
