% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{get_comments}
\alias{get_comments}
\title{get_comments}
\usage{
get_comments(
  ids,
  endpoint = "confirmation",
  billomatApiKey = billomatApiKey,
  billomatID = billomatID
)
}
\arguments{
\item{ids}{a vector of ids numbers}

\item{endpoint}{is the name of the endpoint}
}
\description{
this function pulls all the comments for a list of given ids and endpoint
}
