% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contracts_to_bill.R
\name{get_confirmations_2_bill}
\alias{get_confirmations_2_bill}
\title{get_confirmations_2_bill}
\usage{
get_confirmations_2_bill(
  df_reale_starts,
  confirmations,
  invoices,
  confirmation_items,
  confirmation_number_bill_date,
  billomatDB,
  encryption_key_db = keys$billomat[1],
  create_delayed_7_days = FALSE
)
}
\arguments{
\item{df_reale_starts}{The df which contains the real running times, given the campaigns started}

\item{confirmations}{The df with the confirmations}

\item{invoices}{The df with the invoices}

\item{confirmation_items}{The df with confirmation items}

\item{confirmation_number_bill_date}{The df from asana that contains the number of bills per confirmation}

\item{billomatDB}{The billomat DB location}

\item{encryption_key_db}{The key to decrypt the billomat db}

\item{create_delayed_7_days}{True when a confirmation should get an invoice when at least 7 days old.
When false the confirmations will be filtered depending on the Starting date of the campaign}
}
\value{
The function returns the all the confirmations that need to get an invoice draft.
Confirmations that include CPC, are split or already have an invoice are excluded. These confirmations
should be handled manually.
}
\description{
get_confirmations_2_bill
}
