% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contracts_to_bill.R
\name{get_cpc_document}
\alias{get_cpc_document}
\title{get_cpc_document}
\usage{
get_cpc_document(
  df_items,
  document_type = "confirmation",
  field = "description"
)
}
\arguments{
\item{df_items}{the df with the items which can contain a cpc product}

\item{document_type}{The type of main document (invoice, confirmation).}

\item{field}{The field the information about the product is stored.}
}
\value{
The function returns all document_ids that contain a CPC product
}
\description{
get_cpc_document
}
