% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_api_addresses.R
\name{get_crm_addresses}
\alias{get_crm_addresses}
\title{get_crm_addresses}
\usage{
get_crm_addresses(headers, df)
}
\arguments{
\item{headers}{API headers with authentication}

\item{df}{Dataframe with columns: attachable_id, attachable_type}
}
\value{
Dataframe with all addresses including id, atype, street, city, zip, country, state, etc.
}
\description{
Get all addresses for a person or company
Required columns: attachable_id, attachable_type
}
