% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_api_custom_fields.R
\name{get_crm_custom_fields}
\alias{get_crm_custom_fields}
\title{get_crm_custom_fields}
\usage{
get_crm_custom_fields(headers, df)
}
\arguments{
\item{headers}{API headers with authentication}

\item{df}{Dataframe with columns: attachable_id, attachable_type}
}
\value{
Dataframe with all custom fields including id, custom_fields_type_id, name (value), etc.
}
\description{
Get all custom fields for a person or company
Required columns: attachable_id, attachable_type
}
