% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_api_positions.R
\name{get_crm_positions}
\alias{get_crm_positions}
\title{get_crm_positions}
\usage{
get_crm_positions(headers, df)
}
\arguments{
\item{headers}{API headers with authentication}

\item{df}{Dataframe with column: attachable_id}
}
\value{
Dataframe with all positions including position_id, company_id, name, department, etc.
}
\description{
Get all positions for a person or multiple people
Required columns: attachable_id
}
