% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fields_main_document.R
\name{get_deviating_invoice_recipient}
\alias{get_deviating_invoice_recipient}
\title{get_deviating_invoice_recipient
This function takes the billing information provided and extracts the Debitor or deviating recipient}
\usage{
get_deviating_invoice_recipient(Information_bill_df)
}
\arguments{
\item{Information_bill_df}{The dataframe containing the information about billing details from comments and document}

\item{df_positions}{The dataframe with all the positions}
}
\value{
The function returns a df with all deviating Debitors and their document id
}
\description{
get_deviating_invoice_recipient
This function takes the billing information provided and extracts the Debitor or deviating recipient
}
