% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_process_attachments.R
\name{get_expanded_AB_AN_from_CRM}
\alias{get_expanded_AB_AN_from_CRM}
\title{get_expanded_AB_AN_from_CRM}
\usage{
get_expanded_AB_AN_from_CRM(
  confirmations_billomat,
  offers_billomat,
  attachments,
  simplified_protocols,
  companies
)
}
\arguments{
\item{confirmations_billomat}{confirmations from billomat db}

\item{offers_billomat}{offers from billomat db}

\item{attachments}{attachments file from crm main_data_files}

\item{simplified_protocols}{protocols file (simplified) from crm main_data_files}

\item{companies}{companies file from crm main_data_files (prepared with prepare_companies_crm())}
}
\value{
df with all offers and ABs from the CRM Feed
}
\description{
get_expanded_AB_AN_from_CRM
}
