% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_information_billomat_functions.R
\name{get_extracted_information}
\alias{get_extracted_information}
\title{\subsection{extract the information from each note and put it in a long table}{
}}
\usage{
get_extracted_information(df, field)
}
\arguments{
\item{df}{the df which holds the data from which information is extracted}

\item{field}{the name of the field that contains the data}
}
\value{
the data as df which has id and all information
}
\description{
\subsection{extract the information from each note and put it in a long table}{
}
}
