% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_information_billomat_functions.R
\name{get_impressions_and_bonus}
\alias{get_impressions_and_bonus}
\title{get the number of impressions ordered
this function allows to get the impressions from the orders}
\usage{
get_impressions_and_bonus(df, field2extract = "description")
}
\arguments{
\item{df}{the df which holds the data that was extracted from the field}

\item{field2extract}{the field we want to get}
}
\value{
the data as df with the id and the extracted Impressions and Bonus
}
\description{
get the number of impressions ordered
this function allows to get the impressions from the orders
}
