% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_information_billomat_functions.R
\name{get_invoice_information_from_document}
\alias{get_invoice_information_from_document}
\title{get_invoice_information_from_document
This function takes a data frame with comments and turns them into a table containing the reported performance}
\usage{
get_invoice_information_from_document(df_document, field = "intro")
}
\arguments{
\item{df_comments}{The data frame with all the comments downloaded}
}
\value{
The function returns a df with all the reported clicks in the comment
}
\description{
get_invoice_information_from_document
This function takes a data frame with comments and turns them into a table containing the reported performance
}
