% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_information_billomat_functions.R
\name{get_laufzeiten_information}
\alias{get_laufzeiten_information}
\title{extract the information from each note and put it in a long table
create the final dataframe from the provided data -- important, names must match
the data provided is split with a separate symbol}
\usage{
get_laufzeiten_information(df, keep_keys, is_invoice = TRUE)
}
\arguments{
\item{df}{the df which holds the data from which information is extracted}

\item{keep_keys}{the keys from where laufzeit information is extracted}

\item{is_invoice}{logical, setting whether the df contains information on invoices or not}
}
\value{
the data as df with the id and the extracted laufzeit
}
\description{
extract the information from each note and put it in a long table
create the final dataframe from the provided data -- important, names must match
the data provided is split with a separate symbol
}
