% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_api_people.R
\name{get_responsible_person}
\alias{get_responsible_person}
\title{get_responsible_person}
\usage{
get_responsible_person(persons, employees, o_tag_responsible)
}
\arguments{
\item{persons}{Dataframe containing persons with associated tags.}

\item{employees}{Dataframe containing Vertriebler and Tierpool.}

\item{o_tag_responsible}{Dataframe containing O-Tags and zuständig_group (Zuständiger fuer O-Tag)}
}
\value{
persons_to_include A dataframe containing \code{id} and \code{responsible}, which indicates the
responsible person or group for each \code{id}, determined by the tag priority ('W', 'U', 'O').
}
\description{
This function extracts the responsible person for each entity based on the applied tags.
It prioritizes tags that start with 'W', followed by 'U', and finally 'O'.
}
