% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_shiny_data.R
\name{get_shiny_data_path}
\alias{get_shiny_data_path}
\title{get_shiny_data_path}
\usage{
get_shiny_data_path(
  file = NULL,
  path_to_shiny_files = "../../base-data/shiny_files"
)
}
\arguments{
\item{file}{(optional) The path to the file inside the default location}

\item{path_to_shiny_files}{(optional) the path to the folder where all shiny files are stored. The files of this app are stored in path_to_shiny_files/name_of_the_app}
}
\value{
The path to the file
}
\description{
This function returns the filepath of a particular shiny data file
If file is left empty it returns the default location where data files of this app are stored
}
