% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling_data_billomat_db_functions.R
\name{get_tables_billomat}
\alias{get_tables_billomat}
\title{get_tables_billomat}
\usage{
get_tables_billomat(db_table_name = "confirmations", billomatDB = billomatDB)
}
\arguments{
\item{db_table_name}{the name of the table you are interested in}

\item{billomatDB}{the path to the billomat db}
}
\value{
the data from the DB as dataframe
}
\description{
This function loads the table specified from the local Billomat DB.
}
