% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_helpers.R
\name{has_valid_value}
\alias{has_valid_value}
\title{Check if optional field has valid value}
\usage{
has_valid_value(df, field_name, row_index)
}
\arguments{
\item{df}{dataframe}

\item{field_name}{name of the field to check}

\item{row_index}{row index}
}
\value{
TRUE if field exists, is not NA and not empty string
}
\description{
Check if optional field has valid value
}
\keyword{internal}
