% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_export_queue_handling.R
\name{job_is_running}
\alias{job_is_running}
\title{Check if Job is Currently Running}
\usage{
job_is_running(con, job_name)
}
\arguments{
\item{con}{Database connection object (DBI connection)}

\item{job_name}{Character string identifying the job}
}
\value{
Logical. TRUE if job is running, FALSE otherwise
}
\description{
Determines if a job is currently running by checking the most recent event.
A job is considered running if the last event was 'start' or 'heartbeat'.
}
\examples{
\dontrun{
con <- postgres_connect(needed_tables = c("processed.data_job_events"))
if (job_is_running(con, "my_job")) {
  message("Job is still running")
}
}
}
