% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_helpers.R
\name{map_canton_to_code_ch}
\alias{map_canton_to_code_ch}
\title{Map Swiss canton name to canton code}
\usage{
map_canton_to_code_ch(canton_name)
}
\arguments{
\item{canton_name}{Canton name (e.g., "Zürich", "Bern")}
}
\value{
Canton code (e.g., "ZH", "BE")
}
\description{
Map Swiss canton name to canton code
}
\keyword{internal}
