% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_helpers.R
\name{map_country_to_code}
\alias{map_country_to_code}
\title{Map country name to ISO country code}
\usage{
map_country_to_code(country_name)
}
\arguments{
\item{country_name}{Country name in German (e.g., "Deutschland", "Frankreich", "USA")}
}
\value{
ISO country code (e.g., "DE", "FR", "US")
}
\description{
Map country name to ISO country code
}
