% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_helpers.R
\name{map_state_to_code}
\alias{map_state_to_code}
\title{Map state/canton name to code based on country}
\usage{
map_state_to_code(state_name, country_code)
}
\arguments{
\item{state_name}{State or canton name in German}

\item{country_code}{ISO country code (e.g., "DE", "AT", "CH")}
}
\value{
State/canton code
}
\description{
Map state/canton name to code based on country
}
