% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_helpers.R
\name{map_state_to_code_de}
\alias{map_state_to_code_de}
\title{Map German state name to state code}
\usage{
map_state_to_code_de(state_name)
}
\arguments{
\item{state_name}{State name (e.g., "Bayern", "Nordrhein-Westfalen")}
}
\value{
State code (e.g., "BY", "NW")
}
\description{
Map German state name to state code
}
\keyword{internal}
