% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling_data_billomat_db_functions.R
\name{match_invoices_confirmations}
\alias{match_invoices_confirmations}
\title{match_invoices_confirmations}
\usage{
match_invoices_confirmations(invoices_df, confirmations_df)
}
\arguments{
\item{invoices_df}{the df containing all invoices}

\item{confirmations_df}{the df containing all confirmations}
}
\value{
A df with the confirmation_id, invoice_id and confirmation_number.
}
\description{
this function matches all invoices and confirmations passed via params and returns
their ids and confirmation_number. The function is used to match corrected bills to
its original confirmation
}
