% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_process_attachments.R
\name{matching_ABs_Billomat_CRM}
\alias{matching_ABs_Billomat_CRM}
\title{matching_ABs_Billomat_CRM}
\usage{
matching_ABs_Billomat_CRM(
  attachments,
  confirmations_billomat,
  offers_billomat,
  simplified_protocols
)
}
\arguments{
\item{attachments}{all attachments from CRM}

\item{confirmations_billomat}{all confirmations from Billomat (DB)}

\item{offers_billomat}{all confirmations from Billomat (DB)}

\item{simplified_protocols}{all basic informations about CRM protocols}
}
\value{
A data frame with four columns:
\describe{
\item{attachment_created_at}{(chr) creation date attachement (not document)}
\item{document_id}{(chr) document_id in CRM}
\item{type}{(chr) "Offer" or "AB"}
\item{document_number}{(int) number of Offer or AB}
\item{attachable_id}{(int) id of the protocol or task with the attachment}
\item{person_id}{(int) "Offer" or "AB"}
}
}
\description{
matching_ABs_Billomat_CRM
}
