% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_process_attachments.R
\name{matching_table_AB_AN}
\alias{matching_table_AB_AN}
\title{matching_table_AB_AN}
\usage{
matching_table_AB_AN(confirmations_billomat, offers_billomat)
}
\arguments{
\item{confirmations_billomat}{all confirmations from Billomat (DB)}

\item{offers_billomat}{all confirmations from Billomat (DB)}
}
\value{
A data frame with four columns:
\describe{
\item{attachment_attachable_id}{(int) id of the protocol or task with the attachment}
\item{attachment_created_at}{(chr) creation date attachement (not document)}
\item{document_id}{(chr) document_id in CRM}
\item{type}{(chr) "Offer" or "AB"}
}
}
\description{
matching_table_AB_AN
}
