% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabase_api.R
\name{metabase_get_card}
\alias{metabase_get_card}
\title{Vollstaendige Definition einer Card lesen}
\usage{
metabase_get_card(
  card_id,
  api_key,
  base_url = "https://metabase.studyflix.info"
)
}
\arguments{
\item{card_id}{ID der Card.}

\item{api_key}{Metabase-API-Key.}

\item{base_url}{Basis-URL der Metabase-Instanz.}
}
\value{
Liste mit der Card-Definition inklusive dataset_query.
}
\description{
Vollstaendige Definition einer Card lesen
}
