% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabase_api.R
\name{metabase_sanitize_card}
\alias{metabase_sanitize_card}
\title{Card-Definition von nicht-deterministischen Feldern bereinigen}
\usage{
metabase_sanitize_card(card, dynamic_fields = METABASE_DYNAMIC_FIELDS)
}
\arguments{
\item{card}{Liste mit der Card-Definition (Antwort von \code{metabase_get_card}).}

\item{dynamic_fields}{Character-Vektor der zu entfernenden Feldnamen.}
}
\value{
Liste ohne die dynamischen Felder.
}
\description{
Card-Definition von nicht-deterministischen Feldern bereinigen
}
