% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabase_api.R
\name{metabase_update_card}
\alias{metabase_update_card}
\title{Card aktualisieren}
\usage{
metabase_update_card(
  card_id,
  body,
  api_key,
  base_url = "https://metabase.studyflix.info"
)
}
\arguments{
\item{card_id}{ID der Card.}

\item{body}{Liste mit den zu setzenden Feldern.}

\item{api_key}{Metabase-API-Key.}

\item{base_url}{Basis-URL der Metabase-Instanz.}
}
\value{
Liste mit der aktualisierten Card.
}
\description{
Card aktualisieren
}
