% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_shiny_data.R
\name{name_of_app}
\alias{name_of_app}
\title{name_of_app}
\usage{
name_of_app()
}
\value{
the name of the app (last folder of the current working directory)
}
\description{
Returns the name of the currently running shiny app,
based on the current working directory
}
