% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{post_all_invoice_items}
\alias{post_all_invoice_items}
\title{post_all_invoice_items
This function uses the single invoice item fun and  uses it on a}
\usage{
post_all_invoice_items(df_items, billomatApiKey, billomatID)
}
\arguments{
\item{df_items}{the df containing the items that should be part of the respective invoice already posted.
The df must contain the invoice_id}

\item{billomatID}{the billomat ID}

\item{billomat_api_key}{The Billomat API key.}
}
\value{
The function returns a tibble with the posted confirmation_id and invoice_id
}
\description{
post_all_invoice_items
This function uses the single invoice item fun and  uses it on a
}
