% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{post_all_invoices}
\alias{post_all_invoices}
\title{post_all_invoices
This function uses the post single invoice function and purrrs over a list of confirmations
to turn them into invoices. The function returns a tibble with the confirmation_id and invoice_id}
\usage{
post_all_invoices(confirmations2post, billomatID, billomatApiKey)
}
\arguments{
\item{confirmations2post}{a df with all the confirmations that will be posted}

\item{billomatID}{the billomat ID#'}

\item{billomat_api_key}{The Billomat API key.}
}
\description{
post_all_invoices
This function uses the post single invoice function and purrrs over a list of confirmations
to turn them into invoices. The function returns a tibble with the confirmation_id and invoice_id
}
