% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asana_api.R
\name{post_asana_comment}
\alias{post_asana_comment}
\title{post_asana_comment}
\usage{
post_asana_comment(task_gid, comment_text, api_token, logger)
}
\arguments{
\item{task_gid}{The unique identifier (GID) of the Asana task}

\item{comment_text}{The text of the comment to be posted}

\item{api_token}{The Asana personal access token for authentication}

\item{logger}{A log4r logger object for logging messages}
}
\value{
TRUE if the comment was successfully posted, FALSE otherwise
}
\description{
Posts a comment to an Asana task
}
