% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{post_client_value}
\alias{post_client_value}
\title{post_client_value}
\usage{
post_client_value(
  property_id = 18444,
  client_db = df,
  client_ids = list_of_ids,
  billomatApiKey = billomatApiKey,
  billomatID = billomatID,
  value2post = "debitor_jp5",
  endpoint = "client-property-values"
)
}
\arguments{
\item{property_id}{the id for the property id that is to be changed}

\item{client_db}{this is the table with the client db that contains a client id and the value to be posted}

\item{client_ids}{a list of all client_ids that we want to put}

\item{billomatApiKey}{please provide your billomat Api key here}

\item{billomatID}{please provide your billomat ID here}

\item{value2post}{the string with the name of column name that contains the value}

\item{endpoint}{the string with the name of the endpont where you want to post the value}
}
\description{
this function posts values to a client and the respective table
}
