% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{post_comment}
\alias{post_comment}
\title{Posts comments to a resource via Billomat API}
\usage{
post_comment(billomat_id, billomat_api_key, comment, resource, entity_id)
}
\arguments{
\item{billomat_id}{The Billomat account ID.}

\item{billomat_api_key}{The Billomat API key.}

\item{comment}{The comment you want to post.}

\item{resource}{The resource you want to put (e.g., "invoice", "client", etc.). Always resource in singular!}

\item{entity_id}{The entity id of your object. Findable in the URL.}
}
\value{
A list containing the response from the API.
}
\description{
This function posts comments to a selected resource via Billomat API with the desired comment.
}
