% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{post_complete_invoices}
\alias{post_complete_invoices}
\title{post_complete_invoices}
\usage{
post_complete_invoices(
  billomatID,
  billomatApiKey,
  confirmations2post_df,
  confirmations2post_items_df
)
}
\arguments{
\item{billomatID}{the billomat ID}

\item{df_items}{the df containing the items that should be part of the respective invoice already posted.
The df must contain the invoice_id}

\item{billomat_api_key}{The Billomat API key.}
}
\value{
The functions returns a tibble with all posted ids
}
\description{
This function takes the confirmations and items and turns them into an invoice
}
