% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{post_resource}
\alias{post_resource}
\title{Post a resource to Billomat API}
\usage{
post_resource(billomat_id, billomat_api_key, resource, ...)
}
\arguments{
\item{billomat_id}{The Billomat account ID.}

\item{billomat_api_key}{The Billomat API key.}

\item{resource}{The resource you want to post (e.g., "invoice", "client", etc.). Always resource in singular!}

\item{...}{Additional values corresponding to the resource (see Billomat API documentation).}
}
\value{
A list containing the response from the API.
}
\description{
This function posts a resource to the Billomat API with specified values.
}
