% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{post_single_invoice}
\alias{post_single_invoice}
\title{post_single_invoice
post a single invoice to the billomat. Information for the xml has to be passed via xml dict.}
\usage{
post_single_invoice(
  billomatID,
  billomatApiKey,
  endpoint = "invoices",
  client_id,
  confirmation_id,
  invoice_intro,
  reduction,
  invoice_date,
  note = ""
)
}
\arguments{
\item{billomatID}{the billomat ID#'}

\item{endpoint}{The endpoint you want to post (e.g., "invoices", "clients", etc.). Always resource in plural!}

\item{billomat_api_key}{The Billomat API key.}

\item{entity_id}{The entity id of your object. Findable in the URL.}

\item{...}{Additional values corresponding to the resource (see Billomat API documentation).}
}
\value{
returns a list containing the invoice_id, confirmation_id and invoice_date
}
\description{
post_single_invoice
post a single invoice to the billomat. Information for the xml has to be passed via xml dict.
}
