% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{postgres_add_column}
\alias{postgres_add_column}
\title{postgres_add_column}
\usage{
postgres_add_column(con, schema, table, column_name, column_type)
}
\arguments{
\item{con}{The database connection object.}

\item{schema}{The schema where the table is located.}

\item{table}{The name of the table to which the column will be added.}

\item{column_name}{The name of the new column.}

\item{column_type}{The data type of the new column.}
}
\value{
A feedback message in the console indicating success or failure.
}
\description{
Adds a new column to an existing table in PostgreSQL.
}
