% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{postgres_add_metadata}
\alias{postgres_add_metadata}
\title{postgres_add_metadata}
\usage{
postgres_add_metadata(con, key, value)
}
\arguments{
\item{con}{The database connection object.}

\item{key}{The metadata key.}

\item{value}{The metadata value.}
}
\value{
Only a feedback message in the console.
}
\description{
Adds or updates metadata in the PostgreSQL database.
}
