% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{postgres_change_column_type}
\alias{postgres_change_column_type}
\title{postgres_change_column_type}
\usage{
postgres_change_column_type(con, schema, table, column_name, new_column_type)
}
\arguments{
\item{con}{The database connection object.}

\item{schema}{The schema where the table is located.}

\item{table}{The name of the table containing the column.}

\item{column_name}{The name of the column whose data type will be changed.}

\item{new_column_type}{The new data type for the column.}
}
\value{
A feedback message in the console indicating success or failure.
}
\description{
Changes the data type of an existing column in a PostgreSQL table.
}
