% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{postgres_connect}
\alias{postgres_connect}
\title{postgres_connect}
\usage{
postgres_connect(
  postgres_keys = NULL,
  local_pw = NULL,
  ssl_cert_path = "../../metabase-data/postgres/eu-central-1-bundle.pem"
)
}
\arguments{
\item{postgres_keys}{A named list or object with credentials for production use.}

\item{ssl_cert_path}{Path to the SSL certificate for production use.}
}
\value{
A \code{DBI} connection object if successful. Stops with an error otherwise.
}
\description{
Connects to a PostgreSQL database. Uses secure credentials in production.
In interactive mode, connects to local database instead.
}
