% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{postgres_read_metadata}
\alias{postgres_read_metadata}
\title{postgres_read_metadata}
\usage{
postgres_read_metadata(con, key)
}
\arguments{
\item{con}{The database connection object.}

\item{key}{The metadata key to retrieve.}
}
\value{
A data frame with the metadata key and value, or NULL if the key doesn't exist.
}
\description{
Reads metadata from the PostgreSQL database.
}
