% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{postgres_rename_table}
\alias{postgres_rename_table}
\title{postgres_rename_table}
\usage{
postgres_rename_table(con, schema, old_name, new_name)
}
\arguments{
\item{con}{The database connection object.}

\item{schema}{The schema where the table is located.}

\item{old_name}{The current name of the table.}

\item{new_name}{The new name for the table.}
}
\value{
A message indicating success or failure.
}
\description{
Renames an existing table in a PostgreSQL database.
}
