% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_management_functions.R
\name{postgres_setup_standard_triggers}
\alias{postgres_setup_standard_triggers}
\title{Setup Standard Triggers in a PostgreSQL Table}
\usage{
postgres_setup_standard_triggers(
  con,
  schema,
  table,
  soft_delete = FALSE,
  update_timestamp = FALSE
)
}
\arguments{
\item{con}{A \code{DBI} database connection.}

\item{schema}{A string specifying the schema name where the table resides.}

\item{table}{A string specifying the name of the table to attach triggers to.}

\item{soft_delete}{Logical. Whether to create a soft delete trigger. Defaults to \code{FALSE}.}

\item{update_timestamp}{Logical. Whether to create an updated_at trigger. Defaults to \code{FALSE}.}
}
\value{
No return value. Executes SQL statements on the database to define and attach the requested triggers.
}
\description{
Creates and attaches standard triggers to a PostgreSQL table using a DBI connection.
This includes:
\itemize{
\item An \code{update_timestamp} trigger that updates the \code{updated_at} column when \code{first_name} or \code{name} changes.
\item A \code{soft_delete} trigger that prevents actual deletions and instead sets \code{is_deleted = TRUE} and updates \code{updated_at}.
}
}
\examples{
\dontrun{
postgres_setup_standard_triggers(
  con = con,
  schema = "raw",
  table = "msgraph_users",
  soft_delete = TRUE,
  update_timestamp = TRUE
)
}
}
