% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_process_attachments.R
\name{prepare_companies_crm}
\alias{prepare_companies_crm}
\title{prepare_companies_crm}
\usage{
prepare_companies_crm(companies_crm)
}
\arguments{
\item{companies_crm}{companies file from crm main_data_files}
}
\value{
companies file from crm main_data_files (prepared with prepare_companies_crm())
}
\description{
prepare_companies_crm
}
