% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{pull_production_tables}
\alias{pull_production_tables}
\title{Pulls production tables from a remote PostgreSQL database via SSH}
\usage{
pull_production_tables(
  tables,
  target = c("memory", "local_postgres"),
  ssh_key_path = NULL,
  ssl_cert_path = "../../metabase-data/postgres/eu-central-1-bundle.pem",
  remote_user = "application-user",
  remote_host = "shiny.studyflix.info"
)
}
\arguments{
\item{tables}{A character vector of table names to be retrieved from the remote PostgreSQL database.
#' @param target Specifies where to store the retrieved tables: either 'memory' for an in-memory SQLite database
or 'local_postgres' for a local PostgreSQL database.}

\item{ssh_key_path}{The file path to the private SSH key used for connecting to the remote server.}

\item{ssl_cert_path}{The file path to the SSL certificate for PostgreSQL (default is provided).}

\item{remote_user}{The username for the SSH connection.}

\item{remote_host}{The host address of the remote server.}
}
\value{
A database connection object:
\itemize{
\item If the target is 'memory', returns an SQLite connection object to the in-memory database.
\item If the target is 'local_postgres', returns a PostgreSQL connection object to the local database.
}
}
\description{
This function connects to a remote PostgreSQL database via SSH, retrieves the specified tables,
and stores them either in an in-memory SQLite database or a local PostgreSQL database.
}
