% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_balance.R
\name{put_invoice_value}
\alias{put_invoice_value}
\title{put_invoice_value}
\usage{
put_invoice_value(billomat_api_key, invoice_draft_df, field)
}
\arguments{
\item{billomat_api_key}{The Billomat API key}

\item{invoice_draft_df}{The df containing the invoices}

\item{field}{The field (e.g., "label", "intro", etc.) that should be put to Billomat}
}
\description{
This function puts a value in a selected field via Billomat API for each invoice in a df
}
