% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{put_resource_values}
\alias{put_resource_values}
\title{Put values to a resource via Billomat API}
\usage{
put_resource_values(billomat_id, billomat_api_key, resource, entity_id, ...)
}
\arguments{
\item{billomat_id}{The Billomat account ID.}

\item{billomat_api_key}{The Billomat API key.}

\item{resource}{The resource you want to put (e.g., "invoice", "client", etc.). Always resource in singular!}

\item{entity_id}{The entity id of your object. Findable in the URL.}

\item{...}{Additional values corresponding to the resource (see Billomat API documentation).}
}
\value{
A list containing the response from the API.
}
\description{
This function puts values to a selected resource via Billomat API with specified values.
}
