% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_information_billomat_functions.R
\name{read_KeysFromDescription}
\alias{read_KeysFromDescription}
\title{extract the information from each note and put it in a long table
create the final dataframe from the provided data -- important, names must match
the data provided is split with a separate symbol}
\usage{
read_KeysFromDescription(df, sep = sep, EntryFormatCheck = EntryFormatCheck)
}
\arguments{
\item{df}{the df which holds the data from which information is extracted}

\item{sep}{the separator between key and text}
}
\value{
the data as df which has id and all information
}
\description{
extract the information from each note and put it in a long table
create the final dataframe from the provided data -- important, names must match
the data provided is split with a separate symbol
}
