% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_shiny_data.R
\name{read_shiny_data}
\alias{read_shiny_data}
\title{read_shiny_data}
\usage{
read_shiny_data(file, path_to_shiny_files = "../../base-data/shiny_files")
}
\arguments{
\item{file}{The file path relative to path_to_shiny_files/name_of_the_app}
}
\value{
The R object that was stored in the file
}
\description{
This function reads an RDS file in the folder base-data/shiny_files/name_of_the_app into an R object
}
