% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_connect.R
\name{recursive_search}
\alias{recursive_search}
\title{Recursively Search for Table Names in a Nested List}
\usage{
recursive_search(x)
}
\arguments{
\item{x}{Eine Liste oder ein verschachteltes Objekt aus \code{lazy_query}.}
}
\value{
Kein Rückgabewert – füllt globalen \code{result}-Vektor.
}
\description{
Durchsucht rekursiv ein List-Objekt nach \code{table_names} und speichert sie global.
}
\keyword{internal}
