% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_api_contact_details.R
\name{remove_contact_details}
\alias{remove_contact_details}
\title{remove_contact_details}
\usage{
remove_contact_details(headers, df)
}
\arguments{
\item{headers}{the header informations you have to send with your request}

\item{df}{the dataframe which should include the following fields:}
}
\value{
no return values
}
\description{
This function calls the CRM API and starts a delete request for contact details,
attached to companies or persons.
attachable_id - id of the person / company
custom_fields_id - id of the current contact detail
action - optional, has to be value "remove"
attachable_type - "people" or "companies"
field_type - optional, has to be value "contact_details"
}
