% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_api_addresses.R
\name{remove_crm_addresses}
\alias{remove_crm_addresses}
\title{remove_crm_addresses}
\usage{
remove_crm_addresses(headers, df)
}
\arguments{
\item{headers}{API headers with authentication}

\item{df}{Dataframe with address information}
}
\value{
no return values
}
\description{
Remove addresses from people or companies
Required columns: attachable_id, attachable_type, address_id, action, field_type
action must be "remove", field_type must be "address"
}
