% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_crm_functions.R
\name{remove_crm_custom_fields}
\alias{remove_crm_custom_fields}
\title{remove_crm_custom_fields}
\usage{
remove_crm_custom_fields(headers, df)
}
\arguments{
\item{headers}{the header informations you have to send with your request}

\item{df}{the dataframe which should include the following fields:
attachable_id - id of the person / company
custom_fields_id - id of the current custom_field
action - optional, has to be value "remove"
attachable_type - "people" or "companies"
field_name - id of the custom_field
field_type - optional, has to be value "custom_field"}
}
\value{
no return values
}
\description{
This function calls the CRM API and starts an empty request for
custom_fields, attached to companies or persons.
}
