% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asana_api.R
\name{rename_asana_tag}
\alias{rename_asana_tag}
\title{rename_asana_tag}
\usage{
rename_asana_tag(tag_id, new_name, asana_token, logger)
}
\arguments{
\item{tag_id}{The GID of the tag to rename}

\item{new_name}{The new name for the tag}

\item{asana_token}{Personal Access Token (PAT) for Asana API authentication}

\item{logger}{The logger object}
}
\value{
TRUE if renaming was successful, FALSE otherwise
}
\description{
Renames an existing Asana tag
}
