% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_product.R
\name{rename_product}
\alias{rename_product}
\title{rename_product}
\usage{
rename_product(df)
}
\arguments{
\item{df}{the data frame with the product names}
}
\value{
the df with the products short categories
}
\description{
This function renames the products so that they are consistent
}
