% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{retrieve_and_store_db}
\alias{retrieve_and_store_db}
\title{extract_xml
this function extracts a all entries from an xml}
\usage{
retrieve_and_store_db(
  content,
  billomatApiKey = billomatApiKey,
  billomatID = billomatID,
  billomatDB = billomatDB
)
}
\arguments{
\item{data}{the name of the entry in an xml you are interested in}

\item{list_num}{the index ot the list which contains the data}
}
\value{
the call returns a dataframe which contains the id of the entry and all information in one long df
}
\description{
extract_xml
this function extracts a all entries from an xml
}
