% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{retrieveData}
\alias{retrieveData}
\title{this function carries out the call to get the header
retrieveData
this function}
\usage{
retrieveData(content, per_page)
}
\arguments{
\item{content}{the name of the tables you are interested in, this is one string}

\item{page}{how many entries per page to get}

\item{billomatApiKey}{please provide your billomat Api key here}

\item{billomatID}{please provide your billomat ID here}
}
\value{
the call returns a list with the header information
}
\description{
this function carries out the call to get the header
retrieveData
this function
}
