% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{retrieveData_withCheck}
\alias{retrieveData_withCheck}
\title{retrieveData_withCheck
this function carries out the call to get all the content of the call with validation}
\usage{
retrieveData_withCheck(
  content,
  per_page,
  billomatApiKey = billomatApiKey,
  billomatID = billomatID,
  logger = NULL
)
}
\arguments{
\item{content}{the name of the tables you are interested in, this is one string}

\item{per_page}{how many entries per page to get}

\item{billomatApiKey}{please provide your billomat Api key here}

\item{billomatID}{please provide your billomat ID here}

\item{logger}{a log4r logger object for logging validation results}
}
\value{
the call returns a list with all pages and logs whether the expected row count matches the actual count
}
\description{
retrieveData_withCheck
this function carries out the call to get all the content of the call with validation
}
