% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenue_functions.R
\name{revenue_month}
\alias{revenue_month}
\title{revenue_month}
\usage{
revenue_month(df, current_period, previous_period, revenue)
}
\arguments{
\item{df}{the data frame which must be passed from compute_revenue_month(), there the data is prepared}

\item{current_period, }{the period of interest}

\item{previous_period, }{the period to compare the current period to}

\item{revenue}{the variable containing the revenue}
}
\value{
returns a tibble with grouped revenue streams
}
\description{
Here I compute the revenue per month considering the booking date and the sales person.
Function is used within the compute_revenue_month() function.
This #does not calculate any KPI, only the revenue stream.
this file contains the functions to split revenues in order to later calculate kpis
}
