% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.R
\name{save_downloadable_excel}
\alias{save_downloadable_excel}
\title{save_downloadable_excel}
\usage{
save_downloadable_excel(data, dashboard_nr, file_name)
}
\arguments{
\item{dashboard_nr}{the dashboard number to be included in the file path}

\item{file_name}{the name of the Excel file to be generated}
}
\value{
no return values
}
\description{
This function generates a random password, creates a downloadable Excel file
with the provided filename, and saves it in the specified directory. It also
manages password encryption and updates the password log.
}
