% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_shiny_data.R
\name{save_shiny_data}
\alias{save_shiny_data}
\title{save_shiny_data}
\usage{
save_shiny_data(
  object,
  file,
  path_to_shiny_files = "../../base-data/shiny_files"
)
}
\arguments{
\item{object}{The R object to be saved}

\item{file}{The file path relative to path_to_shiny_files/name_of_the_app}
}
\value{
NULL invisibly
}
\description{
This function saves an R object to an RDS file in the folder base-data/shiny_files/name_of_the_app
If the directory does not exist, the function will create it.
If the file already exists, it will be overwritten
}
