% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_api_companies.R
\name{search_company_by_name}
\alias{search_company_by_name}
\title{search_company_by_name}
\usage{
search_company_by_name(headers, company_name)
}
\arguments{
\item{headers}{API headers with authentication}

\item{company_name}{Name of the company to search for}
}
\value{
List with company information if found, or NULL if not found
}
\description{
Search for company by name in CRM
}
