% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asana_api.R
\name{set_asana_due_date}
\alias{set_asana_due_date}
\title{set_asana_due_date}
\usage{
set_asana_due_date(task_gid, due_date, api_token, logger)
}
\arguments{
\item{task_gid}{The GID of the task to update}

\item{due_date}{The new due date in "YYYY-MM-DD" format}

\item{api_token}{Your Asana API token}

\item{logger}{The logger object}
}
\value{
TRUE if successful, FALSE if not
}
\description{
Sets the due date of an Asana task
}
