% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_balance.R
\name{set_label_invoice}
\alias{set_label_invoice}
\title{set_label_invoice}
\usage{
set_label_invoice(invoices_checked_df, invoice_draft_df)
}
\arguments{
\item{invoices_checked_df}{The df with the checked invoices, this means that all invoices for a contract
have are in sum equal to the contracts total net}

\item{invoice_draft_df}{The df which should be processed to add the label check or remove the label check}
}
\value{
The function returns the invoice_draft_df with the mutated label
}
\description{
set_label_invoice
}
