% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{set_status_confirmations}
\alias{set_status_confirmations}
\title{clear_confirmations}
\usage{
set_status_confirmations(
  confirmation_ids,
  status_to_set,
  billomat_api_key = billomatApiKey,
  billomat_id = billomatID
)
}
\arguments{
\item{confirmation_ids}{an array containing AB numbers}

\item{status_to_set}{a vector to set to either "clear"|"unclear"|"cancel"|"uncancel" depending on what state the confirmation should have}
}
\description{
this function sets the status of confirmations to cleared
}
