% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{set_status_endpoint}
\alias{set_status_endpoint}
\title{set_status_endpoint}
\usage{
set_status_endpoint(
  df,
  endpoint = "confirmations",
  status_to_set,
  billomat_api_key = billomatApiKey,
  billomat_id = billomatID
)
}
\arguments{
\item{status_to_set}{a vector to set to either for confirmations: "clear"|"unclear"|"cancel"|"uncancel"
for the invoice: "complete"}

\item{ids}{an df with the ids for the document (invoice_id,confirmation_id)}
}
\description{
this function sets the status of endpoints to the respective state. Important: for the confirmation / invoice to be completed
you must have a template id -- an invoice / confirmation without any items cannot be completed.
}
