% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenue_functions.R
\name{split_contract_into_days_per_month}
\alias{split_contract_into_days_per_month}
\title{split_contract_into_days_per_month}
\usage{
split_contract_into_days_per_month(date_start, date_end)
}
\arguments{
\item{start_date}{the ending date has to be in the format mm-dd-yyyy}
}
\value{
a tibble
}
\description{
this function takes a starting and ending date and turns it into as many rows as months.
}
