% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_helpers.R
\name{transform_attachable_type}
\alias{transform_attachable_type}
\title{Transform attachable_type to API format}
\usage{
transform_attachable_type(attachable_type)
}
\arguments{
\item{attachable_type}{"people" or "companies"}
}
\description{
Transform attachable_type to API format
}
\keyword{internal}
