% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_crm_functions.R
\name{unnesting_lists}
\alias{unnesting_lists}
\title{unnesting_lists
this function is used to unnest the lists that are part of retrieved table}
\usage{
unnesting_lists(df, listname = listname)
}
\arguments{
\item{df}{the df which contains the list wise variable}

\item{listname}{the lists which should be transformed -- this has to be a list with this name}
}
\value{
the tibble which contains all information, this can be stored in a single vector or lists
}
\description{
unnesting_lists
this function is used to unnest the lists that are part of retrieved table
}
