% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asana_api.R
\name{update_asana_custom_field}
\alias{update_asana_custom_field}
\title{update_asana_custom_field}
\usage{
update_asana_custom_field(
  task_gid,
  custom_field_gid,
  new_value,
  api_token,
  logger
)
}
\arguments{
\item{task_gid}{The GID of the task to update}

\item{custom_field_gid}{The GID of the custom field to update}

\item{new_value}{The new text value to set}

\item{api_token}{Your Asana API token}

\item{logger}{The logger object}
}
\value{
TRUE if successful, FALSE if not
}
\description{
Updates a custom text field in an Asana task
}
