% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asana_api.R
\name{update_asana_custom_field_with_options}
\alias{update_asana_custom_field_with_options}
\title{update_asana_custom_field_with_options}
\usage{
update_asana_custom_field_with_options(
  task_id,
  api_token,
  custom_field_gid,
  option_gids,
  logger
)
}
\arguments{
\item{task_id}{The Asana task ID}

\item{api_token}{The Asana API token}

\item{custom_field_gid}{The GID of the custom field}

\item{option_gids}{A character vector of the GIDs of the options to select}

\item{logger}{The logger object}
}
\value{
TRUE if the custom field was updated successfully, FALSE otherwise
}
\description{
Updates a custom field with optional checkboxes of an Asana task
}
