% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asana_api.R
\name{update_asana_task_assignee}
\alias{update_asana_task_assignee}
\title{update_asana_task_assignee}
\usage{
update_asana_task_assignee(
  task_id,
  assignee_gid,
  workspace_id,
  api_token,
  logger
)
}
\arguments{
\item{task_id}{The Asana task ID}

\item{assignee_gid}{The GID of the new assignee (user)}

\item{workspace_id}{The ID of the workspace where the task resides}

\item{api_token}{The Asana API token}

\item{logger}{The logger object}
}
\value{
TRUE if the assignee was updated successfully, FALSE otherwise
}
\description{
Updates the assignee of an Asana task
}
