% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_api_custom_fields.R
\name{update_crm_custom_fields}
\alias{update_crm_custom_fields}
\title{update_crm_custom_fields}
\usage{
update_crm_custom_fields(headers, df)
}
\arguments{
\item{headers}{the header informations you have to send with your request}

\item{df}{the dataframe which should include the following fields:}
}
\value{
no return values
}
\description{
This function calls the CRM API and updates existing custom_fields,
attached to companies or persons.
attachable_id - id of the person / company
custom_fields_id - id of the custom_field instance to update
action - has to be value "update"
attachable_type - "people" or "companies"
field_name - custom_fields_type_id (the field type ID)
value - new value for the custom field
field_type - has to be value "custom_field"
}
