% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_api_people.R
\name{update_crm_person}
\alias{update_crm_person}
\title{update_crm_person}
\usage{
update_crm_person(headers, df)
}
\arguments{
\item{headers}{the header informations you have to send with your request}

\item{df}{the dataframe which should include the following fields:}
}
\value{
no return values
}
\description{
This function calls the CRM API and updates basic person information
(salutation, first_name, last_name, background).
attachable_id - id of the person (required)
action - has to be value "update"
field_type - has to be value "person"
salutation - optional
first_name - optional
last_name - optional (API field: 'name')
background - optional (description/notes)
}
