% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{upsert_data_to_postgres}
\alias{upsert_data_to_postgres}
\title{upsert_data_to_postgres}
\usage{
upsert_data_to_postgres(connection, schema, table, data)
}
\arguments{
\item{connection}{your database connection}

\item{schema}{the schema of the table in the database}

\item{table}{the table in the database}

\item{data}{the data to be inserted}
}
\value{
Only Feedback Message in Console
}
\description{
upsert_data_to_postgres
}
