% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_helpers.R
\name{validate_attachable_id}
\alias{validate_attachable_id}
\title{Validate attachable_id column}
\usage{
validate_attachable_id(df)
}
\arguments{
\item{df}{dataframe to validate}
}
\description{
Validate attachable_id column
}
\keyword{internal}
