% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_helpers.R
\name{validate_attachable_type}
\alias{validate_attachable_type}
\title{Validate attachable_type column}
\usage{
validate_attachable_type(df)
}
\arguments{
\item{df}{dataframe to validate}
}
\description{
Validate attachable_type column
}
\keyword{internal}
