% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_helpers.R
\name{validate_atype}
\alias{validate_atype}
\title{Validate atype column (for contact details)}
\usage{
validate_atype(df)
}
\arguments{
\item{df}{dataframe to validate}
}
\description{
Validate atype column (for contact details)
}
\keyword{internal}
