% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_helpers.R
\name{validate_custom_fields_id}
\alias{validate_custom_fields_id}
\title{Validate custom_fields_id column}
\usage{
validate_custom_fields_id(df)
}
\arguments{
\item{df}{dataframe to validate}
}
\description{
Validate custom_fields_id column
}
\keyword{internal}
