% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_helpers.R
\name{validate_field_name}
\alias{validate_field_name}
\title{Validate field_name column (for tags)}
\usage{
validate_field_name(df)
}
\arguments{
\item{df}{dataframe to validate}
}
\description{
Validate field_name column (for tags)
}
\keyword{internal}
