% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_helpers.R
\name{validate_not_empty}
\alias{validate_not_empty}
\title{Validate that a column is not empty}
\usage{
validate_not_empty(df, column_name)
}
\arguments{
\item{df}{dataframe to validate}

\item{column_name}{name of the column to validate}
}
\description{
Validate that a column is not empty
}
\keyword{internal}
