% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_helpers.R
\name{validate_positive_id}
\alias{validate_positive_id}
\title{Validate any positive ID column}
\usage{
validate_positive_id(df, id_column_name, id_label = NULL)
}
\arguments{
\item{df}{dataframe to validate}

\item{id_column_name}{name of the ID column to validate}

\item{id_label}{descriptive label for error messages}
}
\description{
Validate any positive ID column
}
\keyword{internal}
