% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_helpers.R
\name{validate_value}
\alias{validate_value}
\title{Validate value column (for custom fields)}
\usage{
validate_value(df)
}
\arguments{
\item{df}{dataframe to validate}
}
\description{
Validate value column (for custom fields)
}
\keyword{internal}
