% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_shiny_data.R
\name{write_data_to_console}
\alias{write_data_to_console}
\title{write_data_to_console}
\usage{
write_data_to_console(df)
}
\arguments{
\item{df}{The R object (normally a data frame) to be processed}
}
\value{
The structure of the input is printed to the console
}
\description{
This function writes the structure of a data frame to the console.
In you want to hardcode the data into a module, you can copy the output of this function.
}
