% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling_data_billomat_db_functions.R
\name{write_db2csv}
\alias{write_db2csv}
\title{write_db2csv}
\usage{
write_db2csv(
  table2process = "invoices",
  billomatDB = billomatDB,
  encryption_key_db = encryption_key_db,
  path_for_export = getwd()
)
}
\arguments{
\item{table2process}{the name of the tables you are interested in, this can be one or a vector}

\item{path_for_export}{the path where you want to store the exported file - default is the working directory}
}
\value{
the data will be stored in the local folder
}
\description{
this function exports the data to a csv
}
