% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenue_functions.R
\name{yearly_nrr_for_month}
\alias{yearly_nrr_for_month}
\title{yearly_nrr_for_month}
\usage{
yearly_nrr_for_month(
  df,
  current_period,
  previous_period,
  revenue,
  focus_company,
  focus_year
)
}
\arguments{
\item{df}{date frame containing the months and revenue streams}

\item{current_period, }{the period of interest}

\item{previous_period, }{the period to compare the current period to}

\item{revenue}{the variable containing the revenue}

\item{focus_company}{whether an overall (FALSE) or company specific (TRUE) NRR should be calculated}

\item{focus_year}{whether the focus of the KPI is the year (TRUE) or a monthly number (FALSE)}
}
\value{
returns a tibble with a calculated NRR
}
\description{
This is a helper function where I compute the nrr and dbner, the nrr and dbner can either be for the company or overall.
The function is called within the compute_nrr_for_month() which also passes the data
}
