% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_api_custom_fields.R
\name{add_crm_custom_fields}
\alias{add_crm_custom_fields}
\title{add_crm_custom_fields}
\usage{
add_crm_custom_fields(headers, df)
}
\arguments{
\item{headers}{the header informations you have to send with your request}

\item{df}{the dataframe which should include the following fields:}
}
\value{
Tibble with created custom_field data (id, custom_fields_type_id, name, attachable_id, etc.) or NULL if no rows processed
}
\description{
This function calls the CRM API and starts an post request for
custom_fields, attached to companies or persons.
attachable_id - id of the person / company
custom_fields_id - optional, id of the current custom_field
action - optional, has to be value "add"
attachable_type - "people" or "companies"
field_name - id of the custom_field
field_type - optional, has to be value "custom_field"
}
