% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_authentication.R
\name{authentication_bonus_db}
\alias{authentication_bonus_db}
\title{authentication_bonus_db}
\usage{
authentication_bonus_db(args)
}
\arguments{
\item{args}{Additional Input Parameter, only needed through FlowForce Job}

\item{return_keys}{optional, vector with already acquired keys}
}
\value{
authentication key in vector
}
\description{
This function executes the Bonus DB authentication process.
It can handle manual password inputs as well as Flow Force args Inputs.
}
