% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_authentication.R
\name{authentication_cleverreach}
\alias{authentication_cleverreach}
\title{authentication_cleverReach}
\usage{
authentication_cleverreach(args)
}
\arguments{
\item{args}{Zusätzlicher Eingabeparameter, nur erforderlich bei FlowForce-Jobs}
}
\value{
Authentifizierungs-Token als Zeichenkette
}
\description{
Diese Funktion führt den Authentifizierungsprozess für CleverReach-RESTAPI durch.
Sie kann sowohl manuelle Passwort-Eingaben als auch FlowForce-Argumente verarbeiten.
}
