% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_authentication.R
\name{authentication_crm}
\alias{authentication_crm}
\title{authentication_crm}
\usage{
authentication_crm(args, path_api_key, return_keys = c())
}
\arguments{
\item{args}{Additional Input Parameter, only needed through FlowForce Job}

\item{path_api_key}{path to api_key decription key}

\item{return_keys}{optional, vector with already acquired keys}
}
\value{
authentication key in vector
}
\description{
This function executes the CRM authentication process.
It can handle manual password inputs as well as Flow Force args Inputs.
}
