% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_authentication.R
\name{authentication_github}
\alias{authentication_github}
\title{authentication_github}
\usage{
authentication_github(args)
}
\arguments{
\item{args}{Additional input parameter, only needed through FlowForce Job}
}
\value{
GitHub Personal Access Token as String
}
\description{
This function handles the key decryption for the GitHub API authentication.
It supports manual decryption key input as well as FlowForce arguments.
}
