% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_authentication.R
\name{authentication_Google_Analytics}
\alias{authentication_Google_Analytics}
\title{authentication_Google_Analytics}
\usage{
authentication_Google_Analytics(args)
}
\arguments{
\item{args}{Additional Input Parameter, only needed through FlowForce Job}
}
\value{
no return values
}
\description{
This function executes the Google Analytics authentication process.
It can handle manual password inputs as well as Flow Force args Inputs.
}
