% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_authentication.R
\name{authentication_Google_BigQuery_GA4}
\alias{authentication_Google_BigQuery_GA4}
\title{authentication_Google_BigQuery_GA4}
\usage{
authentication_Google_BigQuery_GA4(args)
}
\arguments{
\item{args}{Additional Input Parameter, only needed through FlowForce Job}
}
\value{
no return values
}
\description{
This function executes the GA4 BigQuery authentication process for the
bigquery@ga4studyflix.iam.gserviceaccount.com service account.
It decrypts the encrypted key file, authenticates via googleAuthR and bigrquery,
verifies the connection, and deletes the decrypted file afterwards.
It can handle manual password inputs as well as Flow Force args Inputs.
}
