% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_authentication.R
\name{authentication_metabase}
\alias{authentication_metabase}
\title{authentication_metabase}
\usage{
authentication_metabase(args)
}
\arguments{
\item{args}{Additional input parameter, only needed through FlowForce Job}
}
\value{
Metabase API Key as String
}
\description{
This function handles the key decryption for the Metabase API authentication.
It supports manual decryption key input as well as FlowForce arguments.
}
