% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_authentication.R
\name{authentication_msgraph_delegated}
\alias{authentication_msgraph_delegated}
\title{authentication_msgraph_delegated}
\usage{
authentication_msgraph_delegated(args)
}
\arguments{
\item{args}{FlowForce-Decryption-Key.}
}
\value{
Named list(client_secret, store_key).
}
\description{
Decryptet Delegated-App-Secret und Store-Key des Service-Account-Wegs.
}
