% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_authentication.R
\name{authentication_msgraph_scoped_app}
\alias{authentication_msgraph_scoped_app}
\title{authentication_msgraph_scoped_app}
\usage{
authentication_msgraph_scoped_app(args)
}
\arguments{
\item{args}{FlowForce-Decryption-Key.}
}
\value{
App-Secret als String.
}
\description{
Decryptet das Client-Secret der gescopten app-only-MSGraph-App (neuer Weg).
}
