% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_authentication.R
\name{authentication_openai_admin}
\alias{authentication_openai_admin}
\title{authentication_openai_admin}
\usage{
authentication_openai_admin(args)
}
\arguments{
\item{args}{Additional input parameter, only needed through FlowForce Job}
}
\value{
OpenAI Admin API Key as String
}
\description{
This function handles the key decryption for the OpenAI Admin API authentication.
It supports manual decryption key input as well as FlowForce arguments.
}
