% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_authentication.R
\name{authentication_openrouter}
\alias{authentication_openrouter}
\title{authentication_openrouter}
\usage{
authentication_openrouter(args)
}
\arguments{
\item{args}{Additional input parameter, only needed through FlowForce Job}
}
\value{
Openrouter API Key as String
}
\description{
This function handles the key decryption for the Gemini API authentication.
It supports manual decryption key input as well as FlowForce arguments.
}
