% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_authentication.R
\name{authentication_postgresql}
\alias{authentication_postgresql}
\title{authentication_postgresql}
\usage{
authentication_postgresql(args)
}
\arguments{
\item{args}{Additional input parameter, only needed through FlowForce Job}
}
\value{
PostgreSQL DB Key as String
}
\description{
This function handles the key encryption for a PostgreSQL database authentication.
It supports manual password input as well as FlowForce arguments.
}
