% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_authentication.R
\name{authentication_process}
\alias{authentication_process}
\title{authentication_process}
\usage{
authentication_process(
  needed_services = c("billomat", "crm", "google sheet", "asana"),
  args
)
}
\arguments{
\item{needed_services}{the authentication services you need as vector. Currently available: "Billomat", "CRM", "Google Sheet" & "Asana"}

\item{args}{Additional Input Parameter, only needed through FlowForce Job}
}
\value{
authentication keys as vector
}
\description{
This function executes the requested authentication processes.
It can handle manual password inputs as well as Flow Force args Inputs.
}
