% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Billomatics-package.R
\docType{package}
\name{Billomatics-package}
\alias{Billomatics}
\alias{Billomatics-package}
\title{Billomatics: Download and Wrangle Billomat Data}
\description{
The package allows to download and wrangle billomat data and crm data into a df to work with.
}
\author{
\strong{Maintainer}: Johannes Leder \email{johannes.leder@gmail.com} (\href{https://orcid.org/0000-0002-6413-4513}{ORCID})

}
\keyword{internal}
