% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling_data_billomat_db_functions.R
\name{clear_confirmations}
\alias{clear_confirmations}
\title{clear_confirmations}
\usage{
clear_confirmations(
  confirmation_ids,
  billomatApiKey = billomatApiKey,
  billomatID = billomatID
)
}
\arguments{
\item{ab_numbers}{an array containing AB numbers}
}
\description{
this function sets the status of confirmations to cleared
}
