% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fields_main_document.R
\name{create_invoice_recipient}
\alias{create_invoice_recipient}
\title{create_invoice_recipient
This function takes the billing information provided and extracts the Debitor or deviating recipient}
\usage{
create_invoice_recipient(df_positions, Information_bill_df)
}
\arguments{
\item{df_positions}{The dataframe with all the positions}

\item{Information_bill_df}{The dataframe containing the information about billing details from comments and document}
}
\value{
The function returns a df with all deviating Debitors and their document id
}
\description{
create_invoice_recipient
This function takes the billing information provided and extracts the Debitor or deviating recipient
}
