% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{establish_ssh_connection}
\alias{establish_ssh_connection}
\title{Aufbau einer SSH-Verbindung mit Keyfile und Passphrase}
\usage{
establish_ssh_connection(ssh_key_path, remote_user, remote_host)
}
\arguments{
\item{ssh_key_path}{Pfad zum privaten SSH-Key (z. B. \verb{~/.ssh/id_rsa}).}

\item{remote_user}{Benutzername für den Remote-Login.}

\item{remote_host}{Adresse des Remote-Hosts (z. B. \code{server.example.com}).}
}
\value{
Ein aktives SSH-Verbindungsobjekt vom Typ \code{ssh::ssh_session}.
}
\description{
Diese Funktion stellt eine SSH-Verbindung zu einem Remote-Host her. Dabei wird ein privater SSH-Schlüssel
verwendet, dessen Pfad übergeben wird. Die zugehörige Passphrase wird interaktiv abgefragt (via \code{getPass()}).
}
\details{
\itemize{
\item Die Funktion verwendet das Paket \code{ssh}, um eine Verbindung zum Remote-Server aufzubauen.
\item Vor dem Verbindungsaufbau wird geprüft, ob der angegebene SSH-Key existiert.
\item Tritt ein Fehler beim Verbindungsaufbau auf, wird dieser abgefangen und als Fehlermeldung ausgegeben.
}
}
\examples{
\dontrun{
  ssh_session <- establish_ssh_connection(
    ssh_key_path = "~/.ssh/id_rsa",
    remote_user = "mein_nutzer",
    remote_host = "remote.studyflix.de"
  )
}

}
\seealso{
\code{\link[ssh:ssh]{ssh::ssh_connect()}}
}
