% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_APIkey.R
\name{get_billomatApiKey_server}
\alias{get_billomatApiKey_server}
\title{get access to Billomat on server}
\usage{
get_billomatApiKey_server()
}
\arguments{
\item{location_key}{this is the location where the key is found as txt file,
just put it above in the main folder and call it with location_key <- "../key"}
}
\value{
the call returns the decrypted key
}
\description{
this function authenticates at billomat. You will need to enter your created password
}
