% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_functions.R
\name{get_central_station_companies}
\alias{get_central_station_companies}
\title{get_central_station_companies}
\usage{
get_central_station_companies(api_key, positions = TRUE, pages = 2000)
}
\arguments{
\item{api_key}{the api key you have to provide}

\item{positions}{True if the company positions (and person_ids) should also get exported}

\item{pages}{the setting to set a maximum number of pages to download, with 2000 pages as default (one page includes 250 entries)}
}
\value{
the tibble which contains all information, this can be stored in a single vector or lists
}
\description{
this function calls the crm api and downloads the general informations about all companies.
}
