% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_billomat.R
\name{get_comments}
\alias{get_comments}
\title{get_comments}
\usage{
get_comments(
  confirmation_ids,
  billomatApiKey = billomatApiKey,
  billomatID = billomatID
)
}
\arguments{
\item{confirmation_ids}{a vector of ids numbers}
}
\description{
this function pulls all the comments for a list of given ids
}
