% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_information_billomat_functions.R
\name{get_invoice_information_from_document}
\alias{get_invoice_information_from_document}
\title{get_invoice_information_from_document
This function takes a data frame with comments and turns them into a table containing the reported performance}
\usage{
get_invoice_information_from_document(df_document, field = "intro")
}
\arguments{
\item{df_document}{The data frame with all the documents intro}

\item{field}{For the confirmation the information is contained in the "intro". For invoices in "note". #'}
}
\value{
The function returns a df key - value with all the fields that are found in the intro
}
\description{
get_invoice_information_from_document
This function takes a data frame with comments and turns them into a table containing the reported performance
}
