% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_export_queue_handling.R
\name{job_start}
\alias{job_start}
\title{Start a Job Event}
\usage{
job_start(con, job_name)
}
\arguments{
\item{con}{Database connection object (DBI connection)}

\item{job_name}{Character string identifying the job}
}
\value{
Number of rows affected (from DBI::dbExecute)
}
\description{
Records a 'start' event for a job in the data_job_events table.
}
\examples{
\dontrun{
con <- postgres_connect(needed_tables = c("processed.data_job_events"))
job_start(con, "my_export_job")
}
}
