% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabase_api.R
\docType{data}
\name{METABASE_DYNAMIC_FIELDS}
\alias{METABASE_DYNAMIC_FIELDS}
\title{Nicht-deterministische Card-Felder}
\format{
An object of class \code{character} of length 9.
}
\usage{
METABASE_DYNAMIC_FIELDS
}
\description{
Felder, die sich bei jeder Ausfuehrung aendern und deshalb aus exportierten
Card-Definitionen entfernt werden, damit Git-Diffs nur echte Aenderungen an
Query und Visualisierung zeigen.
}
\keyword{datasets}
