% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabase_api.R
\name{metabase_get_cards}
\alias{metabase_get_cards}
\title{Uebersicht aller Cards lesen}
\usage{
metabase_get_cards(api_key, base_url = "https://metabase.studyflix.info")
}
\arguments{
\item{api_key}{Metabase-API-Key.}

\item{base_url}{Basis-URL der Metabase-Instanz.}
}
\value{
Liste der Cards.
}
\description{
Uebersicht aller Cards lesen
}
