% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabase_api.R
\name{metabase_get_collection_items}
\alias{metabase_get_collection_items}
\title{Inhalte einer Collection lesen}
\usage{
metabase_get_collection_items(
  collection_id,
  api_key,
  base_url = "https://metabase.studyflix.info",
  models = NULL
)
}
\arguments{
\item{collection_id}{ID der Collection.}

\item{api_key}{Metabase-API-Key.}

\item{base_url}{Basis-URL der Metabase-Instanz.}

\item{models}{Optionaler Filter, z.B. "card" oder "dataset".}
}
\value{
Liste mit den Items der Collection.
}
\description{
Inhalte einer Collection lesen
}
