% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabase_api.R
\name{metabase_get_tables}
\alias{metabase_get_tables}
\title{Tabellen-Metadaten lesen}
\usage{
metabase_get_tables(api_key, base_url = "https://metabase.studyflix.info")
}
\arguments{
\item{api_key}{Metabase-API-Key.}

\item{base_url}{Basis-URL der Metabase-Instanz.}
}
\value{
Liste der Tabellen.
}
\description{
Wird gebraucht, um in GUI-Fragen (MBQL) die numerische source-table-ID in
einen lesbaren Tabellennamen aufzuloesen.
}
