% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{postgres_connect}
\alias{postgres_connect}
\title{postgres_connect}
\usage{
postgres_connect(
  keys,
  ssl_cert_path = "../../metabase-data/postgres/eu-central-1-bundle.pem"
)
}
\arguments{
\item{keys}{A named list or object with elements: postgresql[\link{1}] = password, [\link{2}] = user, [\link{3}] = dbname, [\link{4}] = host, [\link{5}] = port.}

\item{ssl_cert_path}{Path to the SSL certificate. Default assumes common relative path.}
}
\value{
A \code{DBI} connection object if successful. Stops with an error otherwise.
}
\description{
Connects to a PostgreSQL database using provided credentials and gives feedback.
}
