% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{postgres_drop_table}
\alias{postgres_drop_table}
\title{postgres_drop_table}
\usage{
postgres_drop_table(con, schema, table)
}
\arguments{
\item{con}{The database connection object.}

\item{schema}{The schema where the table is located.}

\item{table}{The name of the table to be dropped.}
}
\value{
A feedback message in the console indicating success or failure.
}
\description{
Drops a table from a PostgreSQL database.
}
