% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{postgres_list_tables}
\alias{postgres_list_tables}
\title{postgres_list_tables}
\usage{
postgres_list_tables(con, schema = NULL)
}
\arguments{
\item{con}{The database connection object.}

\item{schema}{Optional. The schema from which to list the tables. If NULL, lists from all schemas.}
}
\value{
A data frame with schema and table names.
}
\description{
Lists all tables in a specified schema, or all schemas if none is specified.
}
