% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{postgres_rename_table}
\alias{postgres_rename_table}
\title{postgres_rename_table}
\usage{
postgres_rename_table(con, old_name, new_name, schema = "raw")

postgres_rename_table(con, old_name, new_name, schema = "raw")
}
\arguments{
\item{con}{The database connection object.}

\item{old_name}{The current name of the table.}

\item{new_name}{The new name for the table.}

\item{schema}{The schema where the table is located (default is "raw").}
}
\value{
Only a feedback message in the console.

Only a feedback message in the console.
}
\description{
Renames an existing table in a PostgreSQL database.

Renames an existing table in a PostgreSQL database.
}
