% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{postgres_upsert_data}
\alias{postgres_upsert_data}
\title{postgres_upsert_data}
\usage{
postgres_upsert_data(connection, schema, table, data)
}
\arguments{
\item{connection}{your database connection}

\item{schema}{the schema of the table in the database}

\item{table}{the table in the database}

\item{data}{the data to be inserted}
}
\value{
Only Feedback Message in Console
}
\description{
postgres_upsert_data
}
