% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling_data_billomat_db_functions.R
\name{read_most_recent_data}
\alias{read_most_recent_data}
\title{read_most_recent_data}
\usage{
read_most_recent_data(
  location,
  filetype = "RDS",
  name_starts_with = "",
  sheet = 1
)
}
\arguments{
\item{location}{the directors you want to read the most recent file from}

\item{filetype}{the file type you want to read (xlsx,RDS,csv)}

\item{name_starts_with}{filter files by beginning of name}
}
\value{
the most recent data table
}
\description{
this function reads the latest RDS, csv or xlsx file in a certain directory
}
