% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_api_companies.R
\name{update_crm_company}
\alias{update_crm_company}
\title{update_crm_company}
\usage{
update_crm_company(headers, df)
}
\arguments{
\item{headers}{the header informations you have to send with your request}

\item{df}{the dataframe which should include the following fields:}
}
\value{
no return values
}
\description{
This function calls the CRM API and updates basic company information
(name and potentially other fields).
attachable_id - id of the company (required)
action - has to be value "update"
field_type - has to be value "company"
name - optional (company name)
}
